Function to add new equations to the equation table (then used in the get_biomass function)

This function creates S3 objects of class "numeric".

add_equation(taxa, allometry, coords, minDBH, maxDBH, sampleSize,
  unitDBH = "cm", unitOutput = "kg", inputVar = "DBH",
  outputVar = "Total aboveground biomass")

Arguments

taxa
character string or vector specifying the taxon (or taxa) for which the allometry has been calibrated

allometry
a character string with the allometric equation

coords
a vector or matrix of coordinates (longitude, latitude) of the calibration data

minDBH
numerical value, minimum DBH for which the equation is valid (in cm)

maxDBH
numerical value, maximum DBH for which the equation is valid (in cm)

sampleSize
number of measurements with which the allometry was calibrated

unitDBH
character string with unit of DBH in the equation (either cm, mm or inch). Default is cm.

unitOutput
character string with unit of equation output (either g, kg, Mg or lbs if the output is a mass, or m if the output is a height).

inputVar
independent variable(s) needed in the allometry. Default is DBH, other option is DBH, H.

outputVar
dependent variable estimated by the allometry. Default is Total aboveground biomass.

Value

A new equation table including all default equations and the additional equation (equation_id = new)

Examples

new_equations = add_equation(
  taxa = "Faga",
  allometry = "exp(-2+log(dbh)*2.5)",
  coords = c(-0.07, 46.11),
  minDBH = 5,
  maxDBH = 50,
  sampleSize = 50
)