Tables of allometric equations and associated metadata for ForestGEO sites.

A compilation of best available allometry equations to calculate tree above-ground biomass (AGB) per species at ForestGEO sites:

equations: Table of allometric equations.

equations_metadata: Table of metadata of equations.

equations

equations_metadata

Format

A data frame with 168 rows and 22 variables:

# FIXME: Replace with real data.
[,1]	Ozone	 numeric	 Ozone (ppb)
[,2]	Solar.R	 numeric	 Solar R (lang)
[,3]	Wind	 numeric	 Wind (mph)
[,4]	Temp	 numeric	 Temperature (degrees F)
[,5]	Month	 numeric	 Month (1--12)
[,6]	Day	 numeric	 Day of month (1--31)

Details

equation_id: Random, unique equation-identifier.

equation_form: FIXME.

See also