Join allodb tables: equations, sitespecies, and sites_info.

master() has all column types set as characters to preserve different kinds of missing values.

master_tidy() has column types and missing optimized for computation.

master()

master_tidy()

Value

A dataframe.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(purrr)

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

master()

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> # A tibble: 780 x 43
#>    ref_id equation_id equation_allome… equation_form dependent_varia…
#>    <chr>  <chr>       <chr>            <chr>         <chr>           
#>  1 jenki… 2060ea      10^(1.1891+1.41… 10^(a+b*(log… Total abovegrou…
#>  2 jenki… 2060ea      10^(1.1891+1.41… 10^(a+b*(log… Total abovegrou…
#>  3 jenki… a4d879      10^(1.2315+1.63… 10^(a+b*(log… Total abovegrou…
#>  4 jenki… a4d879      10^(1.2315+1.63… 10^(a+b*(log… Total abovegrou…
#>  5 jenki… c59e03      exp(7.217+1.514… exp(a+b*log(… Stem biomass (w…
#>  6 jenki… c59e03      exp(7.217+1.514… exp(a+b*log(… Stem biomass (w…
#>  7 jenki… c59e03      exp(7.217+1.514… exp(a+b*log(… Stem biomass (w…
#>  8 jenki… c59e03      exp(7.217+1.514… exp(a+b*log(… Stem biomass (w…
#>  9 jenki… c59e03      exp(7.217+1.514… exp(a+b*log(… Stem biomass (w…
#> 10 jenki… c59e03      exp(7.217+1.514… exp(a+b*log(… Stem biomass (w…
#> # … with 770 more rows, and 38 more variables: independent_variable <chr>,
#> #   allometry_specificity <chr>, geographic_area <chr>, dbh_min_cm <chr>,
#> #   dbh_max_cm <chr>, sample_size <chr>, dbh_units_original <chr>,
#> #   biomass_units_original <chr>, allometry_development_method <chr>,
#> #   regression_model <chr>, other_equations_tested <chr>, log_biomass <chr>,
#> #   bias_corrected <chr>, bias_correction_factor <chr>,
#> #   notes_fitting_model <chr>, original_data_availability <chr>, warning <chr>,
#> #   site <chr>, family <chr>, species <chr>, species_code <chr>,
#> #   life_form <chr>, equation_group <chr>, equation_taxa <chr>,
#> #   notes_on_species <chr>, wsg_id <chr>, wsg_specificity <chr>, id <chr>,
#> #   Site <chr>, lat <chr>, long <chr>, UTM_Zone <chr>, UTM_X <chr>,
#> #   UTM_Y <chr>, intertropical <chr>, size.ha <chr>, E <chr>,
#> #   wsg.site.name <chr>

# Columns are of type "character"
master() %>%
  purrr::map_chr(typeof) %>%
  unique()

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> [1] "character"

# This preserves preserve different kinds of NAs
unique(grep("N", master()$dbh_min_cm, value = TRUE))

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> [1] "NRA"

allodb::missing_values_metadata

#> # A tibble: 4 x 3
#>   Code  Definition         Description                                          
#>   <chr> <chr>              <chr>                                                
#> 1 <NA>  Not Applicable     Data does not apply to that particular case          
#> 2 NAC   Not Acquired       Information may be available but has not been acquir…
#> 3 NRA   Not Readily Avail… Information was not readily available to the authors…
#> 4 NI    No Information     No information available in original publication     

# For computation, use `master_tidy()`, which has column types and missing
# values already optimized for computation on the data
master_tidy() %>%
  map_chr(typeof) %>%
  unique()

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> [1] "character" "double"    "integer"