Equation references and associated metadata

references: A data frame listing all references used in equation table.

references_metadata: Metadata for reference table.

references

references_metadata

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 57 rows and 6 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 7 rows and 4 columns.

Details

Bibliographical information for sourced equations. Links to the equations table by ref_id.

See also

Examples

#preview the datasets
print(head(references))

#> # A tibble: 6 × 6
#>   ref_id                  ref_doi ref_author ref_year ref_title `References ful…
#>   <chr>                   <chr>   <chr>      <chr>    <chr>     <chr>           
#> 1 ali_2015_abef           dx.doi… Ali        2015     Allometr… "Ali A.,\xa0Xu …
#> 2 barney_1977_bdac        doi.or… Barney     1977     Biomass … "Barney, R. Van…
#> 3 baskerville_1966_dmpi   doi.or… Baskervil… 1966     Dry-Matt… "Baskerville, G…
#> 4 bohn_2014_ocai          doi.or… Bohn       2014     Of clima… "Bohn F, Frak K…
#> 5 bond-lamberty_2002_aabb doi.or… Bond-lamb… 2002     Abovegro… "Bond-Lamberty,…
#> 6 bridge_1979_fpom        NRA     Bridge     1979     Fuelwood… "Bridge J. 1979…

print(head(references_metadata))

#> # A tibble: 6 × 4
#>   Column Field       Description                                   Colum_type   
#>   <chr>  <chr>       <chr>                                         <chr>        
#> 1 1 / A  ref_id      Unique reference identification number to ou… numeric      
#> 2 2 / B  ref_doi     Publication DOI (Digital object identifier)   character (s…
#> 3 3 / C  ref_author  Last name of first author of a cited publica… character (s…
#> 4 4 / D  ref_year    Year of publication                           numeric      
#> 5 5 / E  ref_title   Title of publication                          character (s…
#> 6 6 / F  ref_journal Journal, book, report where published         character (s…