Create objects of class allodb.

This function creates S3 objects of class "allodb". Compared with the data directly delivered in the allodb package, objects of class "allodb" have two main differences:

The column types are not all text but the types that is most useful for computations.

Missing values are represented in a single way: with NA.

set_type(x)

Arguments

x
A dataframe from allodb.

Value

A dataframe subclass "allodb".

Details

The datasets directly delivered by allodb are optimized for delivering the greatest amount of information in the most robust and flexible way. All values are stored as text. This allows, for example, different to store different representation of missing values. The trade-off is that text values don't allow direct computation. First they may need to be converted to other types of data, such as integers.

See also

Examples

set_type(allodb::equations)

#> # A tibble: 175 x 22
#>    ref_id equation_id equation_allome… equation_form dependent_varia…
#>  * <chr>  <chr>       <chr>            <chr>         <chr>           
#>  1 jenki… 2060ea      10^(1.1891+1.41… 10^(a+b*(log… Total abovegrou…
#>  2 jenki… a4d879      10^(1.2315+1.63… 10^(a+b*(log… Total abovegrou…
#>  3 jenki… c59e03      exp(7.217+1.514… exp(a+b*log(… Stem biomass (w…
#>  4 jenki… 96c0af      10^(2.5368+1.31… 10^(a+b*(log… Branches (live,…
#>  5 jenki… 529234      10^(2.0865+0.94… 10^(a+b*(log… Foliage total   
#>  6 jenki… ae65ed      exp(-2.48+2.483… exp(a+b*log(… Total abovegrou…
#>  7 marti… 9c4cc9      10^(-1.326+2.76… 10^(a+b*(log… Total abovegrou…
#>  8 chojn… 7f7777      exp(-2.5095+2.5… exp(a+b*log(… Total abovegrou…
#>  9 jenki… cf733d      exp(5.67+1.97*l… exp(a+b*log(… Total abovegrou…
#> 10 chojn… f08fff      exp(-2.2118+2.4… exp(a+b*log(… Total abovegrou…
#> # … with 165 more rows, and 17 more variables: independent_variable <chr>,
#> #   allometry_specificity <chr>, geographic_area <chr>, dbh_min_cm <dbl>,
#> #   dbh_max_cm <dbl>, sample_size <int>, dbh_units_original <chr>,
#> #   biomass_units_original <chr>, allometry_development_method <chr>,
#> #   regression_model <chr>, other_equations_tested <chr>, log_biomass <chr>,
#> #   bias_corrected <chr>, bias_correction_factor <chr>,
#> #   notes_fitting_model <chr>, original_data_availability <chr>, warning <chr>

class(set_type(allodb::equations))

#> [1] "allodb"      "spec_tbl_df" "tbl_df"      "tbl"         "data.frame"