A dummy table listing sites and species in ForestGEO's network.

This table relates the table of allometric equations by species with the table of allometric equations by site.

site_spp

Format

A data frame with 12 rows and 2 variables:

* `spp`: A character vector of species' names -- in lowercase.

* `site`: A character vector giving site names -- in lowercase.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

site_spp

#> # A tibble: 12 x 2
#>        site    spp
#>       <chr>  <chr>
#>  1      bci swars1
#>  2      bci hybapr
#>  3      bci aegipa
#>  4      bci beilpe
#>  5      bci faraoc
#>  6      bci tet2pa
#>  7 yosemite   pila
#>  8 yosemite   abco
#>  9 yosemite   cade
#> 10 yosemite   conu
#> 11 yosemite   prvi
#> 12 yosemite   quke

full_join(
  full_join(site_spp, site_eqn),
  spp_eqn
)

#> Joining, by = "site"

#> Joining, by = "spp"

#> # A tibble: 12 x 4
#>        site    spp site_eqn spp_eqn
#>       <chr>  <chr>   <list>  <list>
#>  1      bci swars1    <fun>   <fun>
#>  2      bci hybapr    <fun>   <fun>
#>  3      bci aegipa    <fun>   <fun>
#>  4      bci beilpe    <fun>   <fun>
#>  5      bci faraoc    <fun>   <fun>
#>  6      bci tet2pa    <fun>   <fun>
#>  7 yosemite   pila    <fun>   <fun>
#>  8 yosemite   abco    <fun>   <fun>
#>  9 yosemite   cade    <fun>   <fun>
#> 10 yosemite   conu    <fun>   <fun>
#> 11 yosemite   prvi    <fun>   <fun>
#> 12 yosemite   quke    <fun>   <fun>