Function to compute aboveground biomass (or other variables) from allometric equations.

This function creates S3 objects of class "numeric".

get_biomass(dbh, h = NULL, genus = rep(NA, length(dbh)),
  species = NULL, shrub = NULL, coords, new_equations = NULL,
  var = c("Total aboveground biomass", "Whole tree (above stump)"),
  add_weight = FALSE, use_height_allom = TRUE, wna = 0.1,
  wsteep = 3, w95 = 500)

Arguments

dbh
A numerical vector containing tree diameter at breast height (dbh) measurements, in cm.

h
A numerical vector (same length as dbh) containing tree height measurements, in m. Default is NULL, when no measurement is available.

genus
A character vector (same length as dbh), containing the genus (e.g. "Quercus") of each tree. Default is NULL, when no identification is available.

species
A character vector (same length as dbh), containing the species (e.g. "rubra") of each tree. Default is NULL, when no identification is available.

shrub
A logical vector (same length as dbh): is the observation a shrub? Default is NULL (no information), in which case allometric equations designed for shrubs will be used only for species recorded as shrubs in ForestGEO sites.

coords
A numerical vector of length 2 with longitude and latitude (if all trees were measured in the same location) or a matrix with 2 numerical columns giving the coordinates of each tree. Default is NULL when no information is available.

new_equations
Optional. An equation table created with the add_equation() function.

var
What dependent variable(s) should be provided in the output? Default is Total aboveground biomass and Whole tree (above stump), other possible values are: Bark biomass, Branches (dead), Branches (live), Branches total (live, dead), Foliage total, Height, Leaves, Stem (wood only), Stem biomass, Stem biomass (with bark), Stem biomass (without bark), Whole tree (above and belowground). Be aware that only a few equations exist for those other variables, so estimated values might not be very accurate.

add_weight
Should the relative weight given to each equation in the equations data frame be added to the output? Default is FALSE.

use_height_allom
A logical value: should the height allometries from Bohn et al (2014) be used in the AGB allometries from Jansen et al (1996)? Default is TRUE.

wna
this parameter is used in the weighting function to determine the dbh-related weight attributed to equations without a specified dbh range. Default is 0.1

wsteep
this parameter controls the steepness of the dbh-related weight in the weighting function. Default is 3.

w95
this parameter is used in the weighting function to determine the value at which the sample-size-related weight reaches 95 value (max=1). Default is 500.

Value

A vector of class "numeric" of the same length as dbh, containing AGB value (in kg) for every stem, or the dependent variable as defined in var.

Details

The function can run into some memory problems when used on large datasets (usually several hundred thousand observations). In that case, see the 2nd example below for how to apply the function to a splitted dataset.

Examples

data(scbi_stem1)
get_biomass(
  dbh = scbi_stem1$dbh,
  genus = scbi_stem1$genus,
  species = scbi_stem1$species,
  coords = c(-78.2, 38.9)
)

#>    [1] 3.088186e-01 4.470658e-01 3.135326e-01 1.651549e-01 1.100910e+00
#>    [6] 8.299956e-01 5.356413e-01 2.815069e-01 7.120594e-01 3.088186e-01
#>   [11] 9.212900e-01 1.880669e-01 2.208547e+02 2.335499e+00 1.813370e-01
#>   [16] 1.254513e+03 2.667219e+02 8.425356e+02 1.147490e+02 1.874227e+02
#>   [21] 2.707365e+03 1.153633e+03 6.317979e+02 1.489847e+01 1.034495e+01
#>   [26] 6.053704e+02 2.244341e+03 2.259057e+01 2.786129e+03 3.712496e+00
#>   [31] 1.329150e+03 8.686168e+02 2.246966e-01 2.161284e-01 1.725669e+00
#>   [36] 1.008766e+00 5.391219e-01 4.068693e-01 2.841744e+00 4.379423e-01
#>   [41] 8.452285e+00 2.804159e+00 3.829376e-01 6.821929e+01 4.065347e+00
#>   [46] 2.210176e+00 1.669959e+00 6.933497e-01 7.968670e-01 9.480912e-01
#>   [51] 1.653795e-01 2.259232e+00 1.328955e+01 9.185709e-01 1.114783e+00
#>   [56] 1.825720e+00 2.729883e+00 1.669101e+01 7.274597e+00 3.541678e-01
#>   [61] 4.971937e-01 1.739762e+00 2.203874e-01 4.637432e-01 1.250196e+00
#>   [66] 9.680736e-01 6.526725e-01 1.291384e+01 1.514942e-01 3.399779e+00
#>   [71] 2.804159e+00 7.790312e-01 6.769070e-01 1.345514e+00 3.948011e-01
#>   [76] 5.559585e+00 9.680736e-01 3.712776e-01 1.215497e+00 1.321909e-01
#>   [81] 4.008096e-01 4.474506e+00 5.249272e-01 1.865315e+01 6.134189e-01
#>   [86] 1.955917e-01 1.509151e+00 1.995974e-01 7.527162e-01 3.972471e+00
#>   [91] 3.972471e+00 1.147450e+01 5.249272e-01 1.995974e-01 1.321909e-01
#>   [96] 3.160049e-01 1.352812e-01 1.352812e-01 1.419684e+00 3.770822e-01
#>  [101] 3.014598e+00 3.295722e+00 1.238567e+00 3.880879e+00 1.955917e-01
#>  [106] 1.481500e-01 2.203874e-01 2.246966e-01 7.424669e+01 1.689780e-01
#>  [111] 7.614295e-01 1.485682e+01 1.522192e+00 4.379423e-01 7.790312e-01
#>  [116] 8.424878e-01 2.860649e+00 2.562554e-01 1.232235e-01 2.242809e+00
#>  [121] 4.008096e-01 3.948011e-01 7.850290e+00 5.419686e+00 3.266583e-01
#>  [126] 3.213066e-01 2.562554e-01 4.572136e-01 2.753794e-01 3.613736e+00
#>  [131] 6.140629e+00 1.514942e-01 2.161284e-01 6.211564e-01 3.003982e-01
#>  [136] 1.037869e+01 6.413412e+00 2.143094e+01 1.008766e+00 2.705242e-01
#>  [141] 3.712776e-01 1.718323e+01 9.463330e+00 3.141127e+01 1.653795e-01
#>  [146] 1.599494e+01 2.393177e+00 2.379236e-01 4.443133e-01 2.514158e+00
#>  [151] 2.119198e-01 1.291512e-01 3.375116e-01 1.181363e+00 3.153284e+00
#>  [156] 2.729883e+00 4.068693e-01 4.596317e+00 1.733397e+00 6.054434e-01
#>  [161] 2.515981e-01 8.480915e+00 2.077616e-01 1.140648e+01 2.657185e-01
#>  [166] 7.867848e-01 5.178417e-01 2.777149e+00 1.352812e-01 1.609474e+00
#>  [171] 1.010037e+01 1.828825e+01 3.288278e+00 2.036541e-01 3.213066e-01
#>  [176] 2.203874e-01 4.105698e+01 2.909737e+00 1.144478e+00 2.681172e+01
#>  [181] 3.160049e-01 4.703111e-01 1.177846e+00 1.144916e+01 4.971532e-01
#>  [186] 1.569360e+00 2.032642e+02 1.080436e+01 3.448682e+01 2.078808e+02
#>  [191] 1.191656e+02 3.807142e+00 1.823052e+01 1.618319e-01 2.106394e+00
#>  [196] 3.113617e+01 1.430471e+01 3.739757e+00 1.147095e-01 3.213066e-01
#>  [201] 4.507320e-01 2.119198e-01 4.316229e-01 1.384222e-01 2.036541e-01
#>  [206] 2.203874e-01 1.480925e+01 7.943183e+00 3.784599e+00 2.461934e+01
#>  [211] 3.888438e-01 5.390116e-01 3.885138e+01 2.379236e-01 3.784599e+00
#>  [216] 9.346837e+00 1.174974e-01 2.609236e+01 1.800785e-01 5.228842e+00
#>  [221] 1.533499e+02 3.027165e+01 2.346393e+01 1.090107e+00 5.254086e+01
#>  [226] 5.039961e-01 3.107529e-01 4.284624e+02 9.125668e+01 1.917473e+02
#>  [231] 4.939071e+00 1.603349e+00 4.027789e+01 4.400915e+02 6.105292e+02
#>  [236] 1.941851e+03 1.475448e+01 4.989401e+01 5.500772e-01 2.383829e+02
#>  [241] 1.217049e+01 4.614258e+01 8.054122e+01 5.317467e+02 1.419306e+01
#>  [246] 3.000274e+01 4.683812e+02 2.003558e+02 4.204356e+02 8.168857e+00
#>  [251] 2.992371e+01 2.041976e+01 1.141041e+03 1.793787e+01 7.911736e+01
#>  [256] 1.091413e+03 9.821154e+01 5.153194e+00 1.563008e+00 1.993820e+00
#>  [261] 1.741899e+01 1.306189e+01 7.370274e+02 1.370238e+00 2.157135e+01
#>  [266] 2.483806e+02 1.517118e+01 2.531778e+02 2.675713e+02 3.462042e+01
#>  [271] 1.008181e+02 3.210778e+02 1.484120e+00 2.016045e+02 1.936947e-01
#>  [276] 1.333452e+00 1.998220e+01 3.290953e+00 4.860171e+00 5.923614e+01
#>  [281] 8.498282e-01 8.831105e+00 3.095590e+02 7.076636e+01 2.262717e+01
#>  [286] 3.374526e+00 3.418218e+02 3.421571e+02 1.595138e-01 1.630578e+00
#>  [291] 4.084305e+01 5.216522e+01 4.615088e+02 9.766950e-01 6.399575e+00
#>  [296] 2.910921e+00 3.194635e+02 1.100974e+02 7.946788e-01 1.423899e-01
#>  [301] 5.916644e+00 9.145382e+02 1.066656e+02 4.737332e+01 2.229395e+01
#>  [306] 4.305300e+01 1.857284e-01 1.226576e+00 1.349768e+01 5.045454e+00
#>  [311] 5.180342e+00 7.223009e+01 2.400773e+00 2.747322e-01 5.355640e-01
#>  [316] 1.384270e+01 2.405858e+02 4.807992e+00 5.282426e+02 3.410745e+01
#>  [321] 2.315782e+00 2.826958e+02 5.626446e+00 5.099154e+00 3.487857e+01
#>  [326] 6.078924e+01 1.269320e+02 7.687150e+01 9.174233e+00 8.389659e+02
#>  [331] 2.251258e+02 2.964208e+02 8.436080e+02 7.384797e+01 1.603349e+00
#>  [336] 2.583377e+01 1.189676e+02 3.610840e+00 9.194338e+02 9.525461e+00
#>  [341] 1.369418e+01 2.364474e+01 8.498282e-01 2.099079e+01 8.850095e+01
#>  [346] 5.541106e+00 1.169436e+00 1.381571e+03 5.484650e+00 6.151443e+01
#>  [351] 2.884469e+02 6.359793e+01 6.399575e+00 9.251581e+00 8.505455e+01
#>  [356] 6.523939e+00 9.868778e-01 4.209123e-01 2.576258e+00 9.071174e-01
#>  [361] 2.040302e+00 3.626390e+02 2.505179e+00 1.328689e+02 1.147095e-01
#>  [366] 4.420201e+01 3.851904e+00 4.903633e-01 7.484369e+01 9.070091e-01
#>  [371] 9.308390e+00 2.469903e-01 8.371633e+00 2.203874e-01 2.802842e-01
#>  [376] 3.371182e+00 2.317624e+00 1.851529e+02 2.334377e+00 9.758878e-01
#>  [381] 6.238718e+00 6.673371e+00 3.716938e+00 2.666871e+01 6.682325e-01
#>  [386] 5.200746e+00 2.515981e-01 2.928756e+00 4.903633e-01 2.205417e+02
#>  [391] 1.055548e+02 2.833974e+02 2.093925e+01 3.550466e+01 1.084034e+03
#>  [396] 5.718576e+01 1.994191e+02 2.434675e+01 2.131694e+01 2.044186e+01
#>  [401] 3.884804e+01 8.851334e+00 5.060386e+01 1.920907e+02 1.601159e+01
#>  [406] 4.728777e+01 2.739719e+01 2.732091e+02 2.949924e+01 3.716938e+00
#>  [411] 2.531651e+01 4.965767e+01 1.420685e+01 7.630253e+00 5.319013e-01
#>  [416] 5.319013e-01 2.260602e-01 4.791984e-01 2.660804e+00 5.596471e+00
#>  [421] 2.962479e-01 2.010041e+00 9.762386e+00 6.499429e+00 6.082380e+01
#>  [426] 7.721199e+00 6.288289e+00 7.321428e+00 2.454366e-01 1.064561e+01
#>  [431] 9.338208e+00 3.066515e+00 1.411524e+01 4.901908e+00 3.013761e+01
#>  [436] 1.287644e+01 1.052241e+01 3.363513e+00 8.962552e+00 7.823182e+00
#>  [441] 6.438667e+00 2.290557e-01 8.535300e+01 5.718593e+01 8.204419e+00
#>  [446] 5.140029e+01 1.048189e+00 1.207580e+01 1.368436e+02 7.705911e+01
#>  [451] 1.106902e+02 5.161573e+01 4.680117e+01 1.358769e+01 3.817195e+00
#>  [456] 3.776391e+01 4.332511e+02 6.299418e+01 1.179425e+02 9.735070e+01
#>  [461] 9.735070e+01 2.792309e+01 2.471138e+01 3.290326e+01 8.128318e+01
#>  [466] 3.489997e+01 6.372958e+01 1.599063e+01 7.085240e+01 3.684304e+01
#>  [471] 1.382430e+01 3.352055e+02 1.847959e+02 3.793328e+03 3.353918e+01
#>  [476] 7.499848e+03 1.096087e+03 6.250448e+01 4.390523e+03 1.939237e+03
#>  [481] 6.818321e+02 4.095377e+02 6.840554e+01 1.691385e+03 1.788145e+02
#>  [486] 4.867325e+02 2.341824e+03 7.622672e+00 1.637124e+03 4.391790e+02
#>  [491] 3.216342e+03 1.558492e+02 7.653471e+01 6.485239e+01 1.817102e+03
#>  [496] 2.237133e+02 4.322453e+01 8.252730e+02 5.530764e+01 2.854369e+01
#>  [501] 6.250448e+01 5.574976e+03 1.141272e+03 5.054312e+01 1.367548e+01
#>  [506] 1.772244e+02 1.256628e+01 3.081996e+01 1.271864e+02 4.592793e+01
#>  [511] 7.906968e+02 1.423854e+03 2.096675e+01 4.558634e+02 2.137816e+03
#>  [516] 4.491533e+01 4.535243e+03 5.317117e+01 1.116003e+00 9.727054e+03
#>  [521] 4.594053e+02 2.497291e-01 6.341334e+01 1.864805e-01 2.167046e-01
#>  [526] 2.715743e+02 7.236711e+01 1.260512e+02 1.729520e+02 3.969295e+01
#>  [531] 1.974951e-01 1.415285e+01 1.336197e+01 4.849984e+01 4.407880e+00
#>  [536] 5.606011e+01 3.454051e-01 5.056584e+02 1.055240e+02 7.464771e+01
#>  [541] 3.948375e+00 2.122409e+02 9.457887e+01 1.991742e-01 3.669074e-01
#>  [546] 6.382555e-01 3.290018e+00 4.046986e-01 3.292946e-01 4.327408e-01
#>  [551] 3.051106e+00 7.802402e-01 2.711014e-01 2.683293e-01 1.119908e+00
#>  [556] 2.860463e+00 1.374513e+00 1.316260e+00 3.158625e-01 6.954013e-01
#>  [561] 5.473983e-01 5.596683e-01 2.252075e+00 3.062708e-01 1.237430e+00
#>  [566] 2.860463e+00 7.028134e-01 1.772584e+00 1.650390e+00 1.039259e+00
#>  [571] 1.025560e+01 8.465370e-01 1.019726e+00 5.849167e-01 2.823725e-01
#>  [576] 2.484616e+00 5.423130e+00 4.002262e-01 3.229244e+00 1.362732e+00
#>  [581] 5.413519e-01 1.637151e+00 1.237430e+00 4.376169e-01 4.138155e-01
#>  [586] 5.006733e-01 2.993112e+00 3.745872e-01 3.258639e-01 2.910847e-01
#>  [591] 5.978957e-01 3.472204e-01 2.655705e-01 1.434400e+00 7.642726e-01
#>  [596] 4.168410e+00 1.351016e+00 2.126203e+01 7.102855e-01 3.191490e-01
#>  [601] 4.192022e+00 7.330617e-01 7.609307e+00 6.880490e-01 1.704023e+00
#>  [606] 5.235654e-01 9.533223e-01 8.523600e+00 6.414300e+00 2.928933e+01
#>  [611] 4.425509e-01 9.349086e-01 1.204612e+00 9.973524e+00 5.506548e+00
#>  [616] 2.094574e+00 8.550982e-01 2.683293e-01 1.653731e+01 9.533223e-01
#>  [621] 6.178084e-01 3.031580e-01 2.655705e-01 3.224824e-01 1.446574e+00
#>  [626] 3.801692e+00 6.883937e+00 7.330617e-01 7.883148e-01 1.362732e+00
#>  [631] 4.733777e-01 1.089192e+00 3.126215e-01 6.880490e-01 8.811487e-01
#>  [636] 3.547646e-01 8.380368e-01 4.680940e-01 4.950970e-01 1.021501e+01
#>  [641] 6.313804e-01 4.376169e-01 5.473983e-01 4.841203e-01 3.224824e-01
#>  [646] 2.317069e+00 1.151195e+00 6.475761e+00 5.978957e-01 1.059044e+00
#>  [651] 1.571959e+00 3.664917e-01 1.800785e-01 5.902989e-01 3.829376e-01
#>  [656] 2.705242e-01 4.698205e+00 1.174974e-01 7.867848e-01 3.329395e+00
#>  [661] 3.107529e-01 3.160049e-01 4.495774e+00 1.416140e-01 1.006309e+00
#>  [666] 1.877333e-01 2.609622e-01 7.867848e-01 8.410421e-01 1.174974e-01
#>  [671] 1.133480e+00 5.461758e-01 2.368100e+00 1.995974e-01 4.703111e-01
#>  [676] 2.246966e-01 1.726273e-01 3.430134e-01 2.453686e+00 4.421248e+00
#>  [681] 2.246966e-01 1.448566e-01 1.583351e-01 2.246966e-01 2.149403e+03
#>  [686] 1.592487e+03 2.787953e+03 3.242160e+03 1.819115e+03 3.144560e+02
#>  [691] 4.235251e+02 4.066369e+02 9.720725e+02 7.865380e+02 3.117390e+02
#>  [696] 5.076512e+03 8.197635e+02 9.758878e-01 1.352812e-01 3.712776e-01
#>  [701] 2.379236e-01 2.077616e-01 3.375116e-01 2.515981e-01 3.430134e-01
#>  [706] 1.448566e-01 1.304277e+01 1.452569e+00 1.955917e-01 2.753794e-01
#>  [711] 2.705242e-01 4.379406e-01 2.902421e-01 1.341045e+00 1.026892e+00
#>  [716] 2.515981e-01 1.448566e-01 6.682325e-01 6.285731e-01 5.039961e-01
#>  [721] 7.092932e-01 3.829376e-01 1.689780e-01 6.682325e-01 5.902989e-01
#>  [726] 1.076294e+01 1.291512e-01 7.431616e-01 2.424321e-01 2.469903e-01
#>  [731] 2.334648e-01 1.955917e-01 1.838805e-01 1.514942e-01 1.448566e-01
#>  [736] 1.189262e+01 2.515981e-01 2.203874e-01 2.119198e-01 1.838805e-01
#>  [741] 1.481500e-01 1.448566e-01 4.572060e-01 2.753794e-01 6.844878e-01
#>  [746] 5.461758e-01 5.039961e-01 4.836264e-01 3.655238e-01 2.952953e-01
#>  [751] 2.609622e-01 1.548893e-01 1.653795e-01 2.609622e-01 1.838805e-01
#>  [756] 7.779450e-01 2.705242e-01 2.562554e-01 5.039961e-01 4.507320e-01
#>  [761] 3.829376e-01 3.655238e-01 3.213066e-01 2.036541e-01 8.689599e-01
#>  [766] 4.316229e-01 5.178417e-01 1.291512e-01 1.291512e-01 1.416140e-01
#>  [771] 2.119198e-01 6.601891e-01 2.657185e-01 1.321909e-01 1.955917e-01
#>  [776] 8.878661e-01 3.213066e-01 2.705242e-01 1.838805e-01 5.533938e-01
#>  [781] 3.712776e-01 2.246966e-01 1.384222e-01 1.111670e+00 6.285731e-01
#>  [786] 5.039961e-01 3.770822e-01 3.160049e-01 2.562554e-01 2.290557e-01
#>  [791] 1.618319e-01 1.384222e-01 1.775908e+00 1.663885e+00 1.636546e+00
#>  [796] 1.090107e+00 4.129804e-01 1.037275e+00 3.375116e-01 6.844878e-01
#>  [801] 1.321909e-01 3.598206e-01 3.485655e-01 1.955917e-01 2.469903e-01
#>  [806] 1.352812e-01 1.174974e-01 8.878661e-01 3.948011e-01 3.485655e-01
#>  [811] 3.430134e-01 3.375116e-01 2.705242e-01 3.874683e+00 1.955917e-01
#>  [816] 1.583351e-01 1.609474e+00 6.682325e-01 3.055507e-01 4.507320e-01
#>  [821] 3.375116e-01 2.036541e-01 2.802842e-01 1.321909e-01 1.100858e+00
#>  [826] 7.092932e-01 3.213066e-01 1.144478e+00 7.691629e-01 6.208080e-01
#>  [831] 5.178417e-01 1.763274e-01 2.753794e-01 4.379406e-01 1.261621e-01
#>  [836] 6.208080e-01 2.119198e-01 5.978437e-01 2.609622e-01 1.514942e-01
#>  [841] 3.320599e-01 1.261621e-01 6.208080e-01 4.129804e-01 4.572060e-01
#>  [846] 4.836264e-01 1.414813e+00 2.077616e-01 2.246966e-01 4.572060e-01
#>  [851] 2.852383e-01 1.726273e-01 1.689780e-01 1.261621e-01 4.129804e-01
#>  [856] 2.515981e-01 6.285731e-01 1.514942e-01 7.261135e-01 7.092932e-01
#>  [861] 4.703111e-01 3.888438e-01 3.430134e-01 2.609622e-01 1.384222e-01
#>  [866] 1.174974e-01 4.253570e-01 3.213066e-01 2.469903e-01 1.352812e-01
#>  [871] 2.562554e-01 2.515981e-01 2.203874e-01 5.108923e-01 2.424321e-01
#>  [876] 2.161284e-01 1.622977e+00 2.657185e-01 1.800785e-01 1.416140e-01
#>  [881] 8.783836e-01 9.859676e-01 3.712776e-01 7.604383e-01 6.926998e-01
#>  [886] 2.609622e-01 5.178417e-01 1.916371e-01 2.161284e-01 1.514942e-01
#>  [891] 2.290557e-01 5.902989e-01 1.448566e-01 1.548893e-01 1.246270e+00
#>  [896] 5.533938e-01 3.829376e-01 1.955917e-01 3.160049e-01 1.548893e-01
#>  [901] 1.203353e-01 2.852383e-01 4.129804e-01 2.852383e-01 2.424321e-01
#>  [906] 1.800785e-01 2.043812e+00 1.733397e+00 1.026892e+00 4.316229e-01
#>  [911] 2.705242e-01 1.916371e-01 1.763274e-01 1.384222e-01 4.191428e-01
#>  [916] 1.916371e-01 1.583351e-01 6.363936e-01 3.888438e-01 1.995974e-01
#>  [921] 1.916371e-01 1.514942e-01 1.448566e-01 1.416140e-01 2.562554e-01
#>  [926] 2.833430e+00 2.012943e+00 2.424321e-01 1.916371e-01 1.352812e-01
#>  [931] 1.122544e+00 8.502893e-01 6.130982e-01 5.753736e-01 5.178417e-01
#>  [936] 4.572060e-01 3.770822e-01 3.375116e-01 3.888438e-01 3.055507e-01
#>  [941] 2.562554e-01 1.514942e-01 3.375116e-01 3.888438e-01 2.705242e-01
#>  [946] 1.402357e+00 1.257895e+00 9.961079e-01 3.948011e-01 2.562554e-01
#>  [951] 2.203874e-01 1.448566e-01 2.515981e-01 2.334648e-01 1.514942e-01
#>  [956] 1.481500e-01 1.763274e-01 3.888438e-01 5.039961e-01 1.427333e+00
#>  [961] 8.318536e-01 3.655238e-01 1.583351e-01 2.657185e-01 2.802842e-01
#>  [966] 2.161284e-01 1.174974e-01 2.379236e-01 1.068791e+00 5.108923e-01
#>  [971] 1.618319e-01 1.583351e-01 2.721550e+00 1.352812e-01 6.522015e-01
#>  [976] 6.130982e-01 5.248447e-01 4.191428e-01 3.598206e-01 3.107529e-01
#>  [981] 2.515981e-01 2.290557e-01 6.705091e+00 3.375116e-01 2.246966e-01
#>  [986] 1.955917e-01 1.726273e-01 1.689780e-01 1.583351e-01 1.448566e-01
#>  [991] 1.481500e-01 3.829376e-01 7.956823e-01 7.517712e-01 1.414813e+00
#>  [996] 3.829376e-01 3.712776e-01 2.515981e-01 6.054434e-01 1.653795e-01
#> [1001] 8.689599e-01 6.844878e-01 6.522015e-01 5.533938e-01 2.290557e-01
#> [1006] 3.770822e-01 2.036541e-01 6.442697e-01 3.655238e-01 8.878661e-01
#> [1011] 4.443103e-01 1.726273e-01 1.481500e-01 1.416140e-01 1.618319e-01
#> [1016] 1.448566e-01 8.595952e-01 1.804585e+00 6.682325e-01 1.321909e-01
#> [1021] 7.009683e-01 4.316229e-01 5.902989e-01 6.285731e-01 3.160049e-01
#> [1026] 1.514942e-01 4.769424e-01 3.213066e-01 1.232235e-01 1.203353e-01
#> [1031] 3.770822e-01 1.291512e-01 5.248447e-01 4.008096e-01 1.618319e-01
#> [1036] 2.562554e-01 2.290557e-01 8.410421e-01 8.046379e-01 1.232235e-01
#> [1041] 4.769424e-01 4.316229e-01 3.160049e-01 3.485655e-01 2.119198e-01
#> [1046] 3.655238e-01 1.877333e-01 1.653795e-01 1.402357e+00 2.469903e-01
#> [1051] 5.178417e-01 4.443103e-01 2.424321e-01 1.838805e-01 1.174974e-01
#> [1056] 5.606661e-01 6.442697e-01 5.319013e-01 1.916371e-01 2.246966e-01
#> [1061] 5.753736e-01 4.836264e-01 1.516798e+00 3.107529e-01 4.971532e-01
#> [1066] 6.363936e-01 1.877333e-01 1.416140e-01 1.111670e+00 8.410421e-01
#> [1071] 7.604383e-01 2.334648e-01 1.147095e-01 2.074961e+00 9.070091e-01
#> [1076] 2.424321e-01 1.548893e-01 1.995974e-01 8.227235e-01 5.533938e-01
#> [1081] 1.174974e-01 1.452569e+00 8.595952e-01 5.828090e-01 3.107529e-01
#> [1086] 1.352812e-01 1.203353e-01 1.174974e-01 1.377639e+00 1.111670e+00
#> [1091] 9.166698e-01 5.753736e-01 3.770822e-01 2.802842e-01 1.583351e-01
#> [1096] 3.948011e-01 2.609622e-01 1.416140e-01 2.246966e-01 1.416140e-01
#> [1101] 1.478065e+00 9.263899e-01 1.133480e+00 3.655238e-01 1.352812e-01
#> [1106] 3.287924e+00 7.431616e-01 2.609622e-01 2.170100e+00 8.046379e-01
#> [1111] 3.770822e-01 3.055507e-01 1.622977e+00 4.703111e-01 3.320599e-01
#> [1116] 1.174974e-01 2.705242e-01 2.657185e-01 2.246966e-01 1.618319e-01
#> [1121] 1.352812e-01 9.859676e-01 9.361695e-01 8.046379e-01 5.606661e-01
#> [1126] 5.533938e-01 4.507320e-01 3.598206e-01 2.802842e-01 4.769424e-01
#> [1131] 2.036541e-01 1.916371e-01 4.572060e-01 4.769424e-01 3.266583e-01
#> [1136] 2.852383e-01 7.604383e-01 1.583351e-01 2.246966e-01 4.316229e-01
#> [1141] 2.334648e-01 1.389966e+00 1.291512e-01 2.077616e-01 1.916371e-01
#> [1146] 1.763274e-01 1.321909e-01 4.129804e-01 1.916371e-01 1.352812e-01
#> [1151] 4.971532e-01 4.068693e-01 4.008096e-01 3.055507e-01 2.379236e-01
#> [1156] 1.653795e-01 1.514942e-01 9.070091e-01 9.559086e-01 2.657185e-01
#> [1161] 1.877333e-01 1.548893e-01 1.352812e-01 1.448566e-01 5.319013e-01
#> [1166] 4.703111e-01 8.595952e-01 1.203353e-01 4.507320e-01 1.291512e-01
#> [1171] 8.227235e-01 3.213066e-01 6.130982e-01 3.948011e-01 2.952953e-01
#> [1176] 1.726273e-01 4.253570e-01 1.147095e-01 4.129804e-01 2.852383e-01
#> [1181] 1.140648e+01 1.838805e-01 5.461758e-01 9.040103e+00 8.974080e-01
#> [1186] 3.829376e-01 3.003982e-01 1.481500e-01 9.559086e-01 7.176750e-01
#> [1191] 1.916371e-01 4.572060e-01 1.763274e-01 1.261621e-01 1.877333e-01
#> [1196] 1.291512e-01 3.266583e-01 4.507320e-01 1.877333e-01 3.213066e-01
#> [1201] 8.318536e-01 1.232235e-01 2.290557e-01 6.682325e-01 3.125182e+00
#> [1206] 6.330380e+00 5.319013e-01 4.443103e-01 4.443103e-01 4.379406e-01
#> [1211] 3.655238e-01 1.618319e-01 1.481500e-01 8.407966e+00 5.753736e-01
#> [1216] 5.390116e-01 4.836264e-01 3.213066e-01 1.804585e+00 1.246270e+00
#> [1221] 8.502893e-01 7.779450e-01 4.191428e-01 2.515981e-01 6.285731e-01
#> [1226] 5.902989e-01 2.379236e-01 4.637323e-01 4.507320e-01 2.609622e-01
#> [1231] 1.416140e-01 6.682325e-01 4.316229e-01 5.902989e-01 1.100858e+00
#> [1236] 5.606661e-01 3.160049e-01 2.902421e-01 2.161284e-01 1.416140e-01
#> [1241] 1.481500e-01 3.712776e-01 2.351202e+00 1.529840e+00 1.481500e-01
#> [1246] 3.213066e-01 2.036541e-01 1.200403e+00 4.836264e-01 1.726273e-01
#> [1251] 5.828090e-01 4.572060e-01 2.203874e-01 8.595952e-01 4.129804e-01
#> [1256] 1.416140e-01 2.424321e-01 1.174974e-01 1.232235e-01 1.232235e-01
#> [1261] 1.838805e-01 1.448566e-01 2.609622e-01 1.321909e-01 1.384222e-01
#> [1266] 1.877333e-01 2.424321e-01 4.379406e-01 1.261621e-01 1.261621e-01
#> [1271] 1.261621e-01 5.606661e-01 1.726273e-01 1.111670e+00 3.055507e-01
#> [1276] 3.888438e-01 1.321909e-01 1.653795e-01 2.036541e-01 1.952040e+00
#> [1281] 2.562554e-01 3.266583e-01 2.609622e-01 1.653795e-01 1.174974e-01
#> [1286] 1.689780e-01 3.770822e-01 2.334648e-01 1.232235e-01 9.559086e-01
#> [1291] 1.026892e+00 2.161284e-01 4.253570e-01 2.379236e-01 2.203874e-01
#> [1296] 1.448566e-01 5.606661e-01 8.318536e-01 5.319013e-01 4.836264e-01
#> [1301] 1.481500e-01 2.802842e-01 3.055507e-01 2.334648e-01 1.514942e-01
#> [1306] 1.877333e-01 2.203874e-01 4.008096e-01 2.290557e-01 1.174974e-01
#> [1311] 2.515981e-01 3.888438e-01 2.119198e-01 7.346090e-01 1.877333e-01
#> [1316] 3.266583e-01 1.726273e-01 1.838805e-01 1.514942e-01 3.003982e-01
#> [1321] 1.147095e-01 6.130982e-01 1.448566e-01 1.352812e-01 2.290557e-01
#> [1326] 7.779450e-01 4.379406e-01 1.416140e-01 5.365879e+00 1.321909e-01
#> [1331] 2.515981e-01 1.200403e+00 2.119198e-01 3.485655e-01 7.956823e-01
#> [1336] 9.070091e-01 1.321909e-01 4.191428e-01 1.838805e-01 1.618319e-01
#> [1341] 2.379236e-01 3.350249e+00 1.402357e+00 1.328975e+00 1.155538e+00
#> [1346] 8.974080e-01 8.595952e-01 3.655238e-01 1.653795e-01 1.174974e-01
#> [1351] 1.916371e-01 1.261621e-01 2.852342e+00 5.533938e-01 1.548893e-01
#> [1356] 2.246966e-01 2.234948e+00 3.948011e-01 1.352812e-01 1.653795e-01
#> [1361] 2.562554e-01 2.290557e-01 2.753794e-01 6.054434e-01 2.802842e-01
#> [1366] 2.902421e-01 6.522015e-01 2.203874e-01 2.036541e-01 3.829376e-01
#> [1371] 8.974080e-01 5.178417e-01 1.838805e-01 1.618319e-01 2.562554e-01
#> [1376] 1.800785e-01 2.705242e-01 1.384222e-01 1.174974e-01 4.316229e-01
#> [1381] 1.147095e-01 2.802842e-01 1.514942e-01 2.753794e-01 1.618319e-01
#> [1386] 1.481500e-01 5.178417e-01 2.902421e-01 2.753794e-01 1.261621e-01
#> [1391] 3.320599e-01 1.291512e-01 5.039961e-01 1.481500e-01 2.290557e-01
#> [1396] 1.384222e-01 2.334648e-01 2.036541e-01 1.321909e-01 1.261621e-01
#> [1401] 1.416140e-01 2.161284e-01 4.129804e-01 2.119198e-01 1.416140e-01
#> [1406] 1.232235e-01 1.726273e-01 5.461758e-01 8.888898e+00 1.481500e-01
#> [1411] 4.068693e-01 3.655238e-01 2.515981e-01 5.461758e-01 1.689780e-01
#> [1416] 1.321909e-01 1.583351e-01 5.319013e-01 2.290557e-01 2.246966e-01
#> [1421] 1.955917e-01 4.316229e-01 3.541678e-01 2.609622e-01 1.955917e-01
#> [1426] 2.705242e-01 2.705242e-01 1.448566e-01 1.384222e-01 3.055507e-01
#> [1431] 2.119198e-01 2.036541e-01 1.174974e-01 2.952953e-01 1.583351e-01
#> [1436] 1.384222e-01 2.077616e-01 1.583351e-01 3.266583e-01 5.533938e-01
#> [1441] 1.583351e-01 1.726273e-01 5.606661e-01 4.637323e-01 1.122544e+00
#> [1446] 5.902989e-01 4.253570e-01 1.416140e-01 2.077616e-01 1.147095e-01
#> [1451] 2.852383e-01 2.424321e-01 1.689780e-01 1.058224e+00 1.548893e-01
#> [1456] 6.522015e-01 5.978437e-01 5.902989e-01 2.203874e-01 1.838805e-01
#> [1461] 7.956823e-01 7.261135e-01 4.379406e-01 1.877333e-01 1.416140e-01
#> [1466] 2.562554e-01 8.502893e-01 6.601891e-01 4.971532e-01 4.769424e-01
#> [1471] 2.657185e-01 2.334648e-01 1.618319e-01 7.867848e-01 8.974080e-01
#> [1476] 1.653795e-01 1.144478e+00 3.375116e-01 1.800785e-01 1.232235e-01
#> [1481] 1.384222e-01 3.430134e-01 2.562554e-01 1.583351e-01 4.971532e-01
#> [1486] 1.955917e-01 3.888438e-01 1.583351e-01 1.448566e-01 4.507320e-01
#> [1491] 1.448566e-01 1.203353e-01 4.008096e-01 7.691629e-01 2.562554e-01
#> [1496] 1.653795e-01 2.609622e-01 2.290557e-01 5.606661e-01 1.261621e-01
#> [1501] 1.800785e-01 2.609622e-01 1.321909e-01 2.852383e-01 2.753794e-01
#> [1506] 7.956823e-01 5.828090e-01 1.726273e-01 9.758878e-01 2.705242e-01
#> [1511] 2.705242e-01 2.515981e-01 1.384222e-01 4.191428e-01 3.266583e-01
#> [1516] 3.888438e-01 9.758878e-01 2.802842e-01 6.054434e-01 6.763320e-01
#> [1521] 1.689780e-01 2.203874e-01 2.290557e-01 1.726273e-01 2.424321e-01
#> [1526] 2.424321e-01 1.800785e-01 3.541678e-01 1.995974e-01 1.763274e-01
#> [1531] 6.897319e+00 1.147095e-01 1.995974e-01 1.995974e-01 2.246966e-01
#> [1536] 3.712776e-01 2.562554e-01 3.003982e-01 2.203874e-01 1.352812e-01
#> [1541] 5.039961e-01 2.852383e-01 1.689780e-01 1.726273e-01 1.726273e-01
#> [1546] 1.384222e-01 2.203874e-01 5.461758e-01 4.068693e-01 4.316229e-01
#> [1551] 2.609622e-01 4.191428e-01 3.770822e-01 1.916371e-01 6.363936e-01
#> [1556] 1.877333e-01 1.800785e-01 2.852383e-01 7.092932e-01 1.689780e-01
#> [1561] 2.657185e-01 1.514942e-01 7.604383e-01 2.562554e-01 1.653795e-01
#> [1566] 1.548893e-01 3.320599e-01 2.379236e-01 8.046379e-01 3.320599e-01
#> [1571] 1.514942e-01 2.753794e-01 8.335394e+00 1.838805e-01 3.655238e-01
#> [1576] 1.955917e-01 2.036541e-01 8.595952e-01 9.758878e-01 2.562554e-01
#> [1581] 1.223211e+00 1.800785e-01 5.828090e-01 3.003982e-01 3.485655e-01
#> [1586] 1.838805e-01 2.609622e-01 1.147095e-01 2.246966e-01 1.321909e-01
#> [1591] 2.952953e-01 5.828090e-01 4.637323e-01 1.653795e-01 2.290557e-01
#> [1596] 4.379406e-01 2.705242e-01 1.232235e-01 2.902421e-01 2.203874e-01
#> [1601] 2.036541e-01 3.430134e-01 3.055507e-01 1.800785e-01 4.507320e-01
#> [1606] 2.036541e-01 5.606661e-01 1.916371e-01 2.138029e+01 1.486223e+01
#> [1611] 1.291512e-01 1.763274e-01 1.916371e-01 1.689780e-01 1.232235e-01
#> [1616] 5.178417e-01 1.291512e-01 2.246966e-01 6.442697e-01 3.829376e-01
#> [1621] 2.952953e-01 3.829376e-01 1.995974e-01 2.657185e-01 6.363936e-01
#> [1626] 2.902421e-01 1.548893e-01 2.562554e-01 1.689780e-01 2.119198e-01
#> [1631] 1.481500e-01 2.119198e-01 4.507320e-01 2.902421e-01 4.443103e-01
#> [1636] 1.955917e-01 1.384222e-01 1.203353e-01 8.136516e-01 1.416140e-01
#> [1641] 1.203353e-01 5.606661e-01 1.321909e-01 8.318536e-01 3.598206e-01
#> [1646] 1.548893e-01 1.352812e-01 1.416140e-01 1.448566e-01 1.384222e-01
#> [1651] 1.232235e-01 1.203353e-01 1.653795e-01 1.416140e-01 3.430134e-01
#> [1656] 4.507320e-01 2.902421e-01 1.877333e-01 1.352812e-01 5.248447e-01
#> [1661] 5.828090e-01 4.572060e-01 1.416140e-01 5.828090e-01 1.838805e-01
#> [1666] 1.726273e-01 4.769424e-01 1.726273e-01 3.598206e-01 1.203353e-01
#> [1671] 1.232235e-01 1.321909e-01 1.291512e-01 1.955917e-01 1.514942e-01
#> [1676] 3.655238e-01 1.514942e-01 1.548893e-01 4.191428e-01 1.416140e-01
#> [1681] 4.971532e-01 2.609622e-01 3.712776e-01 1.174974e-01 1.147095e-01
#> [1686] 1.352812e-01 3.160049e-01 2.952953e-01 5.648045e+00 1.514942e-01
#> [1691] 2.119198e-01 1.232235e-01 2.562554e-01 1.800785e-01 4.191428e-01
#> [1696] 2.424321e-01 1.689780e-01 1.321909e-01 6.484937e+00 3.107529e-01
#> [1701] 1.321909e-01 6.054434e-01 1.877333e-01 3.107529e-01 1.261621e-01
#> [1706] 6.208080e-01 3.430134e-01 6.682325e-01 1.514942e-01 2.246966e-01
#> [1711] 2.161284e-01 3.107529e-01 1.838805e-01 1.653795e-01 2.077616e-01
#> [1716] 1.203353e-01 1.232235e-01 1.877333e-01 1.955917e-01 3.213066e-01
#> [1721] 2.119198e-01 2.424321e-01 1.174974e-01 1.763274e-01 1.147095e-01
#> [1726] 3.888438e-01 1.203353e-01 1.514942e-01 4.379406e-01 1.147095e-01
#> [1731] 4.253570e-01 3.266583e-01 3.485655e-01 6.285731e-01 1.203353e-01
#> [1736] 1.291512e-01 2.161284e-01 1.763274e-01 4.008096e-01 3.003982e-01
#> [1741] 2.609622e-01 2.469903e-01 8.407966e+00 1.402357e+00 2.424321e-01
#> [1746] 2.902421e-01 5.533938e-01 1.481500e-01 3.375116e-01 2.609622e-01
#> [1751] 1.203353e-01 3.266583e-01 1.174974e-01 6.208080e-01 2.119198e-01
#> [1756] 2.203874e-01 1.174974e-01 1.261621e-01 1.321909e-01 4.903633e-01
#> [1761] 1.384222e-01 1.481500e-01 2.379236e-01 2.119198e-01 1.582666e+00
#> [1766] 1.133480e+00 1.384222e-01 1.877333e-01 2.705242e-01 1.877333e-01
#> [1771] 2.952953e-01 4.507320e-01 1.147095e-01 3.107529e-01 2.379236e-01
#> [1776] 1.689780e-01 9.263899e-01 1.232235e-01 1.514942e-01 1.352812e-01
#> [1781] 2.753794e-01 5.606661e-01 2.077616e-01 7.691629e-01 3.598206e-01
#> [1786] 1.416140e-01 6.285731e-01 2.161284e-01 2.077616e-01 2.753794e-01
#> [1791] 1.384222e-01 2.802842e-01 1.174974e-01 1.838805e-01 4.971532e-01
#> [1796] 1.653795e-01 1.548893e-01 1.416140e-01 3.712776e-01 4.443103e-01
#> [1801] 3.829376e-01 5.282923e+00 2.609622e-01 1.761671e+00 1.689780e-01
#> [1806] 9.658682e-01 4.129804e-01 4.129804e-01 1.995974e-01 1.995974e-01
#> [1811] 2.852383e-01 2.379236e-01 2.036541e-01 3.213066e-01 1.321909e-01
#> [1816] 6.208080e-01 1.800785e-01 2.852383e-01 1.481500e-01 2.203874e-01
#> [1821] 3.598206e-01 2.609622e-01 3.541678e-01 1.352812e-01 1.689780e-01
#> [1826] 1.514942e-01 7.867848e-01 1.203353e-01 2.657185e-01 2.562554e-01
#> [1831] 5.390116e-01 1.006309e+00 1.916371e-01 1.174974e-01 3.246768e+00
#> [1836] 1.800785e-01 1.481500e-01 1.174974e-01 1.384222e-01 1.416140e-01
#> [1841] 1.232235e-01 1.352812e-01 1.877333e-01 7.517712e-01 1.653795e-01
#> [1846] 5.533938e-01 2.424321e-01 2.562554e-01 1.305027e+00 1.481500e-01
#> [1851] 1.147095e-01 1.321909e-01 1.291512e-01 7.867848e-01 1.514942e-01
#> [1856] 1.800785e-01 1.321909e-01 1.416140e-01 5.606661e-01 5.828090e-01
#> [1861] 9.070091e-01 4.572060e-01 3.107529e-01 1.416140e-01 1.203353e-01
#> [1866] 3.770822e-01 2.562554e-01 1.653795e-01 3.519932e+00 1.877333e-01
#> [1871] 1.416140e-01 3.655238e-01 2.469903e-01 1.548893e-01 3.375116e-01
#> [1876] 3.829376e-01 1.291512e-01 1.321909e-01 1.481500e-01 1.321909e-01
#> [1881] 1.481500e-01 1.203353e-01 1.352812e-01 2.246966e-01 2.246966e-01
#> [1886] 2.036541e-01 9.658682e-01 4.106960e+00 2.077616e-01 2.753794e-01
#> [1891] 2.802842e-01 1.481500e-01 2.777149e+00 1.800785e-01 2.469903e-01
#> [1896] 3.541678e-01 3.107529e-01 1.291512e-01 4.395585e+01 2.161480e+03
#> [1901] 6.275821e+00 1.818128e+02 8.600541e+02 3.274394e+02 6.710287e+01
#> [1906] 3.006620e+03 4.331408e+03 3.673043e+02 4.552154e+03 3.414316e+03
#> [1911] 1.186741e+03 3.001867e+03 8.790002e+02 4.872791e+02 1.626256e+01
#> [1916] 8.595952e-01 1.855496e+02 3.987035e-01 5.770021e+01 3.348456e+02
#> [1921] 1.871391e+00 2.746280e+02 1.103674e+01 1.660969e+02 2.634460e+01
#> [1926] 2.983057e+00 1.300124e+03 3.738180e+02 2.802842e-01 1.230660e+01
#> [1931] 9.102940e+02 8.354925e+02 1.893514e-01 1.285208e+02 4.925655e-01
#> [1936] 1.030342e+01 5.896935e+02 8.529660e+01 1.441896e+02 1.570997e+00
#> [1941] 6.560811e+00 7.214944e+00 2.653212e+01 1.258250e+03 5.550452e+02
#> [1946] 1.461866e+03 1.865261e-01 2.228737e+01 4.731800e+03 5.161462e-01
#> [1951] 3.950321e+03 4.028676e+02 2.089324e+02 2.616106e+01 1.717887e+02
#> [1956] 3.704755e-01 1.344834e+03 7.502727e+02 4.194356e+02 2.473203e+03
#> [1961] 6.559983e+02 1.304905e+02 1.591862e+01 2.191304e+00 1.010097e+03
#> [1966] 1.776896e+03 2.224328e+03 5.816597e+02 1.660510e-01 8.000530e+00
#> [1971] 3.460950e+02 1.759034e+03 1.902180e+03 2.961556e+02 1.470905e+03
#> [1976] 2.007862e+02 7.282568e+02 2.296811e+03 5.582134e+02 5.212869e+02
#> [1981] 5.659646e+02 1.728918e+03 1.760726e-01 1.465766e+02 1.588837e+03
#> [1986] 3.972730e+00 8.209556e+02 2.046775e+02 2.140187e+03 1.773993e+02
#> [1991] 1.017851e+03 2.380418e+01 9.326459e+01 9.868048e+02 5.211379e+00
#> [1996] 2.977464e+03 1.101528e+00 3.109854e+02 2.647251e+03 6.728245e+03
#> [2001] 1.021451e+01 3.781823e+01 1.057794e+03 1.410448e+02 5.062535e+01
#> [2006] 1.975588e+03 1.079834e+03 2.037402e+03 3.427981e+01 1.689395e-01
#> [2011] 1.155845e+03 1.090408e+00 6.240900e+01 4.875079e+01 4.001532e+02
#> [2016] 3.264951e+00 3.531426e-01 2.655549e+02 9.335527e+00 1.828512e+03
#> [2021] 5.460198e+02 5.928888e+01 3.363780e+03 1.777946e+03 3.314488e+03
#> [2026] 1.212379e+02 3.834242e-01 3.367573e-01 3.367573e-01 6.697253e-01
#> [2031] 3.136565e-01 2.807269e-01 3.709078e-01 1.542962e+00 2.028343e+00
#> [2036] 2.416760e-01 5.986527e-01 2.807269e-01 9.705114e-01 9.038226e-01
#> [2041] 3.075975e-01 2.380062e-01 2.567948e-01 6.623232e-01 2.908119e-01
#> [2046] 2.759902e-01 6.405107e-01 2.759902e-01 2.628546e-01 4.292833e-01
#> [2051] 8.056485e-01 6.405107e-01 2.736934e-01 2.490986e-01 3.920858e-01
#> [2056] 3.367573e-01 2.908119e-01 2.380062e-01 5.986527e-01 2.736934e-01
#> [2061] 4.651948e-01 2.882110e-01 1.146197e+00 2.759902e-01 2.548349e-01
#> [2066] 3.075975e-01 2.670685e-01 2.567948e-01 2.472247e-01 2.989479e-01
#> [2071] 2.435166e-01 3.668623e-01 1.341406e+00 3.403022e-01 2.783342e-01
#> [2076] 3.403022e-01 2.453651e-01 3.439091e-01 5.919049e-01 5.219717e-01
#> [2081] 6.262965e-01 8.492723e-01 1.622977e+00 8.853719e-01 2.416760e-01
#> [2086] 4.292833e-01 2.831697e-01 5.655651e-01 2.587830e-01 2.649426e-01
#> [2091] 2.670685e-01 5.402655e-01 3.199536e-01 3.075975e-01 3.105974e-01
#> [2096] 2.509896e-01 3.750166e-01 2.736934e-01 2.736934e-01 2.628546e-01
#> [2101] 4.984862e-01 3.017733e-01 2.649426e-01 2.435166e-01 3.231925e-01
#> [2106] 3.136565e-01 2.490986e-01 2.416760e-01 8.762468e-01 4.927763e-01
#> [2111] 1.967162e+00 3.105974e-01 3.017733e-01 2.882110e-01 8.228983e-01
#> [2116] 2.649426e-01 2.628546e-01 2.759902e-01 2.548349e-01 9.415279e-01
#> [2121] 3.475782e-01 2.882110e-01 2.908119e-01 2.608021e-01 2.453651e-01
#> [2126] 2.856640e-01 3.046563e-01 3.017733e-01 2.807269e-01 2.692343e-01
#> [2131] 2.934679e-01 2.759902e-01 4.651948e-01 2.759902e-01 5.402655e-01
#> [2136] 3.439091e-01 4.984862e-01 5.042606e-01 2.628546e-01 1.146197e+00
#> [2141] 2.934679e-01 3.264919e-01 7.077218e-01 2.567948e-01 2.548349e-01
#> [2146] 2.587830e-01 3.046563e-01 5.852223e-01 2.692343e-01 5.720525e-01
#> [2151] 2.856640e-01 2.567948e-01 2.472247e-01 2.587830e-01 4.815504e-01
#> [2156] 2.807269e-01 2.587830e-01 4.342208e-01 2.856640e-01 4.651948e-01
#> [2161] 5.280048e-01 3.668623e-01 4.148548e-01 2.736934e-01 2.398403e-01
#> [2166] 2.934679e-01 3.136565e-01 4.101733e-01 4.196003e-01 8.404144e-01
#> [2171] 5.527854e-01 2.908119e-01 3.017733e-01 3.298524e-01 3.231925e-01
#> [2176] 3.046563e-01 2.989479e-01 2.682232e-01 3.203617e-01 1.961798e-01
#> [2181] 2.284303e+00 2.094988e-01 2.018155e-01 2.001282e-01 2.418213e-01
#> [2186] 2.267493e-01 2.079230e-01 2.385460e-01 1.966349e-01 1.961798e-01
#> [2191] 1.960700e-01 4.052879e-01 2.079230e-01 1.971233e-01 1.963051e-01
#> [2196] 1.961798e-01 2.295008e-01 1.982854e-01 2.487540e-01 1.961798e-01
#> [2201] 2.295008e-01 1.963051e-01 9.810028e-02 1.645846e+00 1.579372e+00
#> [2206] 1.413898e+00 8.635123e-01 1.353049e+00 1.654708e-01 8.822387e-01
#> [2211] 1.321909e-01 1.689780e-01 4.191428e-01 2.802842e-01 4.971532e-01
#> [2216] 1.653795e-01 2.515981e-01 2.246966e-01 1.763274e-01 3.375116e-01
#> [2221] 1.203353e-01 1.800785e-01 3.712776e-01 2.609622e-01 1.261621e-01
#> [2226] 3.107529e-01 2.203874e-01 1.448566e-01 1.416140e-01 1.321909e-01
#> [2231] 9.206228e+01 1.358594e+03 2.538812e+01 4.290360e+00 2.310643e-01
#> [2236] 2.266158e-01 2.351597e+00 5.366154e+02 2.581892e+00 6.961871e+00
#> [2241] 1.772266e-01 1.240381e+00 1.969036e-01 3.102193e+01 2.421072e+00
#> [2246] 3.142983e+01 5.277828e+01 6.059422e+01 1.697198e-01 1.849415e-01
#> [2251] 9.190778e+01 2.300289e+00 1.290719e-01 3.637368e+00 4.509811e-01
#> [2256] 5.917477e+00 8.927779e+01 5.760181e+01 5.155672e+01 2.636895e+00
#> [2261] 8.896071e+00 1.644953e+02 4.338242e+01 9.444525e+00 1.362256e+02
#> [2266] 2.711330e+00 3.812658e+00 1.517064e+01 4.455835e+00 1.081139e+00
#> [2271] 1.088807e+01 6.746019e-01 2.919361e-01 5.133151e+00 2.563600e-01
#> [2276] 7.658518e-01 2.121546e-01 3.062288e-01 1.122037e+00 3.744906e+00
#> [2281] 5.849885e-01 1.149271e+01 1.499860e+00 1.410746e+01 4.085149e-01
#> [2286] 5.164652e-01 2.884804e-01

# split dataset to avoid memory over usage
data_split = split(scbi_stem1, cut(1:nrow(scbi_stem1), breaks = 10, labels = FALSE))
agb = lapply(data_split, function(df) get_biomass(dbh=df$dbh,
                                                 genus=df$genus,
                                                 species = df$species,
                                                 coords = c(-78.2, 38.9)))
scbi_stem1$agb = do.call(c, agb)