Function to compute aboveground biomass (or other variables) from allometric equations.

This function creates S3 objects of class "numeric".

get_biomass(
  dbh,
  h = NULL,
  genus = rep(NA, length(dbh)),
  species = NULL,
  shrub = NULL,
  coords,
  new_equations = NULL,
  var = c("Total aboveground biomass", "Whole tree (above stump)"),
  add_weight = FALSE,
  use_height_allom = TRUE,
  wna = 0.1,
  wsteep = 3,
  w95 = 500
)

Arguments

dbh
A numerical vector containing tree diameter at breast height (dbh) measurements, in cm.

h
A numerical vector (same length as dbh) containing tree height measurements, in m. Default is NULL, when no measurement is available.

genus
A character vector (same length as dbh), containing the genus (e.g. "Quercus") of each tree. Default is NULL, when no identification is available.

species
A character vector (same length as dbh), containing the species (e.g. "rubra") of each tree. Default is NULL, when no identification is available.

shrub
A logical vector (same length as dbh): is the observation a shrub? Default is NULL (no information), in which case allometric equations designed for shrubs will be used only for species recorded as shrubs in ForestGEO sites (see data("shrub_species")).

coords
A numerical vector of length 2 with longitude and latitude (if all trees were measured in the same location) or a matrix with 2 numerical columns giving the coordinates of each tree. Default is NULL when no information is available.

new_equations
Optional. An equation table created with the add_equation() function.

var
What dependent variable(s) should be provided in the output? Default is Total aboveground biomass and Whole tree (above stump), other possible values are: Bark biomass, Branches (dead), Branches (live), Branches total (live, dead), Foliage total, Height, Leaves, Stem (wood only), Stem biomass, Stem biomass (with bark), Stem biomass (without bark), Whole tree (above and belowground). Be aware that only a few equations exist for those other variables, so estimated values might not be very accurate.

add_weight
Should the relative weight given to each equation in the equations data frame be added to the output? Default is FALSE.

use_height_allom
A logical value: should the height allometries from Bohn et al (2014) be used in the AGB allometries from Jansen et al (1996)? Default is TRUE.

wna
this parameter is used in the weighting function to determine the dbh-related weight attributed to equations without a specified dbh range. Default is 0.1

wsteep
this parameter controls the steepness of the dbh-related weight in the weighting function. Default is 3.

w95
this parameter is used in the weighting function to determine the value at which the sample-size-related weight reaches 95% of its maximum value (max=1). Default is 500.

Value

A vector of class "numeric" of the same length as dbh, containing AGB value (in kg) for every stem, or the dependent variable as defined in var.

Details

The function can run into some memory problems when used on large datasets (usually several hundred thousand observations). In that case, see the 2nd example below for how to apply the function to a splitted dataset.

Examples

data(scbi_stem1)
get_biomass(
dbh = scbi_stem1$dbh,
genus = scbi_stem1$genus,
species = scbi_stem1$species,
coords = c(-78.2, 38.9)
)

#>    [1]    0.3118052    0.4548249    0.3166681    0.1649523    1.1326884
#>    [6]    0.8519732    0.5466671    0.2836637    0.7296695    0.3118052
#>   [11]    0.9466096    0.1881017  220.3533837    2.4073249    0.1812868
#>   [16] 1246.5560988  266.3091009  839.1431488  114.9714280  187.3569728
#>   [21] 2667.9021553 1147.0247237  629.9676226   15.1540823   10.5699123
#>   [26]  603.6975051 2217.3552707   22.8733653 2744.3507273    3.8317433
#>   [31] 1320.0916014  864.9990308    0.2246966    0.2752643    1.8187963
#>   [36]    1.1041263    0.6258282    0.4869439    2.9176884    0.5198973
#>   [41]    8.3544781    2.8808630    0.4613879   65.3533141    4.1118400
#>   [46]    2.2973460    1.7635936    0.7846330    0.8900799    1.0430353
#>   [51]    0.2123668    2.3456581   13.0037946    1.0132601    1.2105654
#>   [56]    1.9178364    2.8080583   16.2644452    7.2190958    0.4304210
#>   [61]    0.5821498    1.8327547    0.2803223    0.5470904    1.3460474
#>   [66]    1.0631700    0.7429794   12.6432884    0.1940178    3.4633180
#>   [71]    2.8808630    0.8719615    0.7678124    1.4411501    0.4740776
#>   [76]    5.5617054    1.0631700    0.4488724    1.3113748    0.1673704
#>   [81]    0.4804885    4.5095718    0.6110711   18.1431901    0.7026413
#>   [86]    0.2504583    1.6039976    0.2553558    0.8451939    4.0214657
#>   [91]    4.0214657   11.2613647    0.6110711    0.2553558    0.1673704
#>   [96]    0.3888268    0.1717352    0.1717352    1.5150234    0.4551085
#>  [101]    3.0869161    3.3617208    1.3344301    3.9323039    0.2504583
#>  [106]    0.1895023    0.2803223    0.2854131   71.0715924    0.2170255
#>  [111]    0.8540619   14.5068582    1.6169551    0.5198973    0.8719615
#>  [116]    0.9363415    2.9362068    0.3219947    0.1544756    2.3294867
#>  [121]    0.4804885    0.4740776    7.7743392    5.4262365    0.4005067
#>  [126]    0.3946466    0.3219947    0.5402217    0.3436725    3.6720304
#>  [131]    6.1238665    0.1940178    0.2752643    0.7106047    0.3716048
#>  [136]   10.2082983    6.3875412   20.8007881    1.1041263    0.3381985
#>  [141]    0.4488724   16.7358802    9.3278833   30.3338972    0.2123668
#>  [146]   15.5976198    2.4774559    0.3008855    0.5266255    2.5963552
#>  [151]    0.2702390    0.1630379    0.4123486    1.2772390    3.2225467
#>  [156]    2.8080583    0.4068693    4.5963169    1.7333972    0.6054434
#>  [161]    0.2515981    8.4809154    0.2077616   11.4064770    0.2657185
#>  [166]    0.7867848    0.5178417    2.7771489    0.1352812    1.6094737
#>  [171]   10.0457591   18.1390309    3.2858125    0.2067034    0.3241841
#>  [176]    0.2234754   40.5892871    2.9083976    1.1452766   26.5501149
#>  [181]    0.3189020    0.4726046    1.1786015   11.3804212    0.4993494
#>  [186]    1.5697040  199.7445943   10.7424671   34.1166574  204.2651804
#>  [191]  117.3414147    3.8026350   18.0820236    0.1645605    2.1063377
#>  [196]   30.8141583   14.2038663    3.7355451    0.1164752    0.3241841
#>  [201]    0.4530989    0.2149930    0.4340633    0.1407703    0.2067034
#>  [206]    0.2234754   14.7024860    7.9093190    3.7801918   24.3875500
#>  [211]    0.3914540    0.5410644   38.4168662    0.2410169    3.7801918
#>  [216]    9.2997588    0.1193424   25.8406036    0.1829942    5.2164046
#>  [221]  150.8632264   29.9618778   23.2476132    1.0909784   51.8938054
#>  [226]    0.5061682    0.3136690  431.8588588   91.2724691  192.4273727
#>  [231]    4.8481580    1.5623383   40.1039004  443.6390968  616.4821017
#>  [236] 1978.8908219   14.6006227   49.7449912    0.5340375  239.4797059
#>  [241]   12.0278737   45.9827302   80.5077323  536.5368198   14.0414434
#>  [246]   29.8180344  472.3016511  201.1067718  423.7283994    8.0496199
#>  [251]   29.7390259   20.2482772 1156.7921767   17.7730715   79.0776622
#>  [256] 1106.1379140   98.2617181    5.0599702    1.5228265    1.9451684
#>  [261]   17.2557814   12.9151222  745.0454585    1.3341341   21.3972453
#>  [266]  249.5797603   15.0157156  254.4275417  268.9782466   34.4376276
#>  [271]  100.8816587  323.1179126    1.4455846  202.3658695    0.1888325
#>  [276]    1.2981485   19.8117897    3.2208891    4.7701296   59.1179936
#>  [281]    0.8258304    8.7073370  311.4583487   70.6918227   22.4509057
#>  [286]    3.3032544  344.1189112  344.4584584    0.1559010    1.5890115
#>  [291]   40.6702338   52.0232547  465.3379173    0.9495750    6.2943056
#>  [296]    2.8465895  321.4838716  110.2107452    0.7720791    0.1394200
#>  [301]    5.8157821  925.7842771  106.7601327   47.2168985   22.1183476
#>  [306]   42.8852020    0.1811532    1.1936415   13.3489878    4.9533824
#>  [311]    5.0868305   72.1614668    2.3448150    0.2670745    0.5199350
#>  [316]   13.6925368  241.7047363    4.7185328  532.9832537   33.9241811
#>  [321]    2.2613013  284.2752093    5.5283829    5.0065055   34.6960382
#>  [326]   60.6766515  127.1440484   76.8221685    9.0482393  848.8003369
#>  [331]  226.0941982  298.1618208  853.5277324   73.7859703    1.5623383
#>  [336]   25.6523276  119.1316525    3.5362511  930.7728230    9.3972682
#>  [341]   13.5446460   23.4666220    0.8258304   20.8179653   88.5036282
#>  [346]    5.4438897    1.1377951 1402.7140681    5.3879998   61.4044669
#>  [351]  290.0934979   63.4956490    6.2943056    9.1250968   85.0412661
#>  [356]    6.4175822    0.9595125    0.4086350    2.5173264    0.8816941
#>  [361]    1.9907852  365.1968944    2.4474390  133.1182916    0.1409326
#>  [366]   43.5614685    3.9246333    0.5740065   73.7589020    1.0006126
#>  [371]    9.2813898    0.3105455    8.3628377    0.2794299    0.3484104
#>  [376]    3.4508430    2.4092373  182.8443445    2.4258473    1.0701997
#>  [381]    6.2704751    6.6970285    3.7916824   26.3167956    0.7577524
#>  [386]    5.2512509    0.3158495    3.0140859    0.5740065  220.5416918
#>  [391]  105.5547551  283.3973690   20.9392462   35.5046627 1084.0340549
#>  [396]   57.1857602  199.4190874   24.3467472   21.3169433   20.4418640
#>  [401]   38.8480377    8.8513339   50.6038561  192.0906874   16.0115894
#>  [406]   47.2877656   27.3971946  273.2090536   29.4992359    3.7169381
#>  [411]   25.3165144   49.6576704   14.2068462    7.6751897    0.5421744
#>  [416]    0.5421744    0.2231906    0.4687225    2.6092350    5.4925011
#>  [421]    0.2907920    1.9695020    9.5749109    6.3782114   59.1788323
#>  [426]    7.5758608    6.1711498    7.1840777    0.2414671   10.4392425
#>  [431]    9.1596680    3.0080278   13.8311066    4.8108490   29.4352379
#>  [436]   12.6207074   10.3186992    3.2999043    8.7918482    7.6757911
#>  [441]    6.3186262    0.2290557   83.9809463   56.3852867    8.1818956
#>  [446]   50.7030312    1.0945559   11.9817868  134.1996035   75.8661156
#>  [451]  108.7235852   50.9147189   46.1823272   13.4677049    3.8602613
#>  [456]   37.2893564  422.7148626   62.0846185  115.7940282   95.7055412
#>  [461]   95.7055412   27.5925341   24.4260016   32.5012553   80.0000357
#>  [466]   34.4685530   62.8058783   15.8319104   69.7877697   36.7324175
#>  [471]   13.8596375  331.2688133  182.9144939 3692.9560799   33.4549360
#>  [476] 7290.4877134 1079.9442247   62.1578458 4266.4264226 1904.8338209
#>  [481]  672.9985699  404.6798580   67.9987187 1663.0295254  177.0192934
#>  [486]  480.8538835 2296.0586623    7.6716755 1610.0047641  433.9391660
#>  [491] 3139.4579085  154.3756448   76.0419877   64.4820320 1785.7632090
#>  [496]  221.2543628   43.0598274  814.0128600   55.0317257   28.4966659
#>  [501]   62.1578458 5412.7552459 1124.2859115   50.3122708   13.7113573
#>  [506]  175.4518531   12.6058854   30.7564193  126.0930407   45.7392571
#>  [511]  780.0351802 1401.3255590   20.9682380  450.4032276 2098.0558872
#>  [516]   44.7357023 4405.7941370   52.9156502    1.1352940 9475.3782181
#>  [521]  453.8978639    0.2530910   63.0538379    0.1884213    0.2193192
#>  [526]  268.3616493   71.9144183  124.9570316  171.2077784   39.5545276
#>  [531]    0.1996792   14.1813685   13.3933856   48.2843088    4.4488928
#>  [536]   55.7734823    0.3508854  499.3206303  104.6900262   74.1705482
#>  [541]    3.9878332  209.9017867   93.8765133    0.2012581    0.3728453
#>  [546]    0.6222516    3.2137928    0.3957159    0.3234447    0.4227769
#>  [551]    2.9799353    0.7605441    0.2684544    0.2658437    1.0918980
#>  [556]    2.7933636    1.3405100    1.2836159    0.3106868    0.6778861
#>  [561]    0.5339052    0.5458260    2.1982146    0.3016042    1.2066354
#>  [566]    2.7933636    0.6851050    1.7294457    1.6100313    1.0131816
#>  [571]   10.0475660    0.8251748    0.9941189    0.5703668    0.2790636
#>  [576]    2.4256502    5.3038057    0.3914068    3.1542998    1.3290033
#>  [581]    0.5280323    1.5970942    1.2066354    0.4274889    0.4045061
#>  [586]    0.4885507    2.9231755    0.3667491    0.3201820    0.2872689
#>  [591]    0.5829874    0.3405364    0.2632440    1.3990068    0.7449823
#>  [596]    4.0740299    1.3175606   20.8753875    0.6923829    0.3138042
#>  [601]    4.0971617    0.7145705    7.4486515    0.6707260    1.6624413
#>  [606]    0.5107626    0.9293277    8.3462931    6.2759437   28.7836148
#>  [611]    0.4322586    0.9113642    1.1745915    9.7704304    5.3856005
#>  [616]    2.0442051    0.8335228    0.2658437   16.2245414    0.9293277
#>  [621]    0.6023562    0.2986616    0.2632440    0.3169688    1.4108994
#>  [626]    3.7148138    6.7367393    0.7145705    0.7684142    1.3290033
#>  [631]    0.4620932    1.0619156    0.3076152    0.6707260    0.8589272
#>  [636]    0.3477493    0.8168867    0.4569758    0.4831430   10.0076895
#>  [641]    0.6155613    0.4274889    0.5339052    0.4725020    0.3169688
#>  [646]    2.2617759    1.1224399    6.3362412    0.5829874    1.0324902
#>  [651]    1.5333950    0.3589820    0.1800785    0.5902989    0.3829376
#>  [656]    0.2705242    4.6982054    0.1174974    0.7867848    3.3293953
#>  [661]    0.3107529    0.3160049    4.4957743    0.1416140    1.0063087
#>  [666]    0.1877333    0.2609622    0.7867848    0.8410421    0.1174974
#>  [671]    1.1334803    0.5461758    2.3680996    0.1995974    0.4703111
#>  [676]    0.2246966    0.1726273    0.3430134    2.4536857    4.4212484
#>  [681]    0.2246966    0.1448566    0.1583351    0.2246966 2234.4288893
#>  [686] 1646.6575182 2917.2930304 3407.6477298 1884.9656133  320.0503406
#>  [691]  432.1967348  414.8238560  999.3827182  806.9449707  317.2579772
#>  [696] 5396.2460450  841.3699102    0.9579217    0.1346732    0.3661773
#>  [701]    0.2356026    0.2060214    0.3331272    0.2490031    0.3385126
#>  [706]    0.1441296   12.7005288    1.4242198    0.1940726    0.2722979
#>  [711]    0.2675429    0.4314235    0.2868519    1.3151517    1.0078292
#>  [716]    0.2490031    0.1441296    0.6568281    0.6180101    0.4960758
#>  [721]    0.6970172    0.3775896    0.1679044    0.6568281    0.5805476
#>  [726]   10.4870159    0.1286133    0.7301660    0.2400214    0.2444883
#>  [731]    0.2312319    0.1940726    0.1825644    0.1506782    0.1441296
#>  [736]   11.5839692    0.2490031    0.2184090    0.2101020    0.1825644
#>  [741]    0.1473794    0.1441296    0.4502794    0.2722979    0.6727384
#>  [746]    0.5373604    0.4960758    0.4761386    0.3605456    0.2917997
#>  [751]    0.2581769    0.1540257    0.1643615    0.2581769    0.1825644
#>  [756]    0.7642095    0.2675429    0.2535660    0.4960758    0.4439430
#>  [761]    0.3775896    0.3605456    0.3172646    0.2019896    0.8532840
#>  [766]    0.4252400    0.5096275    0.1286133    0.1286133    0.1409285
#>  [771]    0.2101020    0.6489554    0.2628359    0.1316188    0.1940726
#>  [776]    0.8717861    0.3172646    0.2675429    0.1825644    0.5444253
#>  [781]    0.3661773    0.2226350    0.1377764    1.0907766    0.6180101
#>  [786]    0.4960758    0.3718586    0.3120746    0.2535660    0.2269094
#>  [791]    0.1608674    0.1377764    1.7403499    1.6308376    1.6041099
#>  [796]    1.0696801    0.4069938    1.0179884    0.3331272    0.6727384
#>  [801]    0.1316188    0.3549634    0.3439470    0.1940726    0.2444883
#>  [806]    0.1346732    0.1170791    0.8717861    0.3892010    0.3439470
#>  [811]    0.3385126    0.3331272    0.2675429    3.7887796    0.1940726
#>  [816]    0.1574222    1.5776412    0.6568281    0.3018402    0.4439430
#>  [821]    0.3331272    0.2019896    0.2771011    0.1316188    1.0801982
#>  [826]    0.6970172    0.3172646    1.1228738    0.7556143    0.6104097
#>  [831]    0.5096275    0.1751367    0.2722979    0.4314235    0.1256566
#>  [836]    0.6104097    0.2101020    0.5879324    0.2581769    0.1506782
#>  [841]    0.3277909    0.1256566    0.6104097    0.4069938    0.4502794
#>  [846]    0.4761386    1.3872970    0.2060214    0.2226350    0.4502794
#>  [851]    0.2819524    0.1714962    0.1679044    0.1256566    0.4069938
#>  [856]    0.2490031    0.6180101    0.1506782    0.7134802    0.6970172
#>  [861]    0.4631061    0.3833703    0.3385126    0.2581769    0.1377764
#>  [866]    0.1170791    0.4191073    0.3172646    0.2444883    0.1346732
#>  [871]    0.2535660    0.2490031    0.2184090    0.5028256    0.2400214
#>  [876]    0.2142312    1.5908434    0.2628359    0.1788262    0.1409285
#>  [881]    0.8625063    0.9677849    0.3661773    0.7470753    0.6807762
#>  [886]    0.2581769    0.5096275    0.1901877    0.2142312    0.1506782
#>  [891]    0.2269094    0.5805476    0.1441296    0.1540257    1.2224516
#>  [896]    0.5444253    0.3775896    0.1940726    0.3120746    0.1540257
#>  [901]    0.1198895    0.2819524    0.4069938    0.2819524    0.2400214
#>  [906]    0.1788262    2.0021958    1.6987934    1.0078292    0.4252400
#>  [911]    0.2675429    0.1901877    0.1751367    0.1377764    0.4130253
#>  [916]    0.1901877    0.1574222    0.6256648    0.3833703    0.1980066
#>  [921]    0.1901877    0.1506782    0.1441296    0.1409285    0.2535660
#>  [926]    2.7733287    1.9720295    0.2400214    0.1901877    0.1346732
#>  [931]    1.1014153    0.8350120    0.6028634    0.5659389    0.5096275
#>  [936]    0.4502794    0.3718586    0.3331272    0.3833703    0.3018402
#>  [941]    0.2535660    0.1506782    0.3331272    0.3833703    0.2675429
#>  [946]    1.3751156    1.2338222    0.9777072    0.3892010    0.2535660
#>  [951]    0.2184090    0.1441296    0.2490031    0.2312319    0.1506782
#>  [956]    0.1473794    0.1751367    0.3833703    0.4960758    1.3995415
#>  [961]    0.8169697    0.3605456    0.1574222    0.2628359    0.2771011
#>  [966]    0.2142312    0.1170791    0.2356026    1.0488240    0.5028256
#>  [971]    0.1608674    0.1574222    2.6641303    0.1346732    0.6411373
#>  [976]    0.6028634    0.5164819    0.4130253    0.3549634    0.3069331
#>  [981]    0.2490031    0.2269094    6.5438682    0.3331272    0.2226350
#>  [986]    0.1940726    0.1714962    0.1679044    0.1574222    0.1441296
#>  [991]    0.1473794    0.3775896    0.7815692    0.7385925    1.3872970
#>  [996]    0.3775896    0.3661773    0.2490031    0.5953710    0.1643615
#> [1001]    0.8532840    0.6727384    0.6411373    0.5444253    0.2269094
#> [1006]    0.3718586    0.2019896    0.6333738    0.3605456    0.8717861
#> [1011]    0.4376578    0.1714962    0.1473794    0.1409285    0.1608674
#> [1016]    0.1441296    0.8441193    1.7683822    0.6568281    0.1316188
#> [1021]    0.6888691    0.4252400    0.5805476    0.6180101    0.3120746
#> [1026]    0.1506782    0.4695966    0.3172646    0.1227487    0.1198895
#> [1031]    0.3718586    0.1286133    0.5164819    0.3950817    0.1608674
#> [1036]    0.2535660    0.2269094    0.8259622    0.7903341    0.1227487
#> [1041]    0.4695966    0.4252400    0.3120746    0.3439470    0.2101020
#> [1046]    0.3605456    0.1863516    0.1643615    1.3751156    0.2444883
#> [1051]    0.5096275    0.4376578    0.2400214    0.1825644    0.1170791
#> [1056]    0.5515434    0.6333738    0.5233888    0.1901877    0.2226350
#> [1061]    0.5659389    0.4761386    1.4870268    0.3069331    0.4893782
#> [1066]    0.6256648    0.1863516    0.1409285    1.0907766    0.8259622
#> [1071]    0.7470753    0.2312319    0.1143171    2.0326346    0.8905197
#> [1076]    0.2400214    0.1540257    0.1980066    0.8080342    0.5444253
#> [1081]    0.1170791    1.4242198    0.8441193    0.5732166    0.3069331
#> [1086]    0.1346732    0.1198895    0.1170791    1.3509418    1.0907766
#> [1091]    0.8999735    0.5659389    0.3718586    0.2771011    0.1574222
#> [1096]    0.3892010    0.2581769    0.1409285    0.2226350    0.1409285
#> [1101]    1.4491519    0.9094855    1.1121144    0.3605456    0.1346732
#> [1106]    3.2167390    0.7301660    0.2581769    2.1255954    0.7903341
#> [1111]    0.3718586    0.3018402    1.5908434    0.4631061    0.3277909
#> [1116]    0.1170791    0.2675429    0.2628359    0.2226350    0.1608674
#> [1121]    0.1346732    0.9677849    0.9190555    0.7903341    0.5515434
#> [1126]    0.5444253    0.4439430    0.3549634    0.2771011    0.4695966
#> [1131]    0.2019896    0.1901877    0.4502794    0.4695966    0.3225033
#> [1136]    0.2819524    0.7470753    0.1574222    0.2226350    0.4252400
#> [1141]    0.2312319    1.3629973    0.1286133    0.2060214    0.1901877
#> [1146]    0.1751367    0.1316188    0.4069938    0.1901877    0.1346732
#> [1151]    0.4893782    0.4010126    0.3950817    0.3018402    0.2356026
#> [1156]    0.1643615    0.1506782    0.8905197    0.9383714    0.2628359
#> [1161]    0.1863516    0.1540257    0.1346732    0.1441296    0.5233888
#> [1166]    0.4631061    0.8441193    0.1198895    0.4439430    0.1286133
#> [1171]    0.8080342    0.3172646    0.6028634    0.3892010    0.2917997
#> [1176]    0.1714962    0.4191073    0.1143171    0.4069938    0.2819524
#> [1181]   11.1119420    0.1825644    0.5373604    8.8135235    0.8811239
#> [1186]    0.3775896    0.2967958    0.1473794    0.9383714    0.7052209
#> [1191]    0.1901877    0.4502794    0.1751367    0.1256566    0.1863516
#> [1196]    0.1286133    0.3225033    0.4439430    0.1863516    0.3172646
#> [1201]    0.8169697    0.1227487    0.2269094    0.6568281    3.0580002
#> [1206]    6.1794290    0.5233888    0.4376578    0.4376578    0.4314235
#> [1211]    0.3605456    0.1608674    0.1473794    8.1992767    0.5659389
#> [1216]    0.5303482    0.4761386    0.3172646    1.7683822    1.2224516
#> [1221]    0.8350120    0.7642095    0.4130253    0.2490031    0.6180101
#> [1226]    0.5805476    0.2356026    0.4566671    0.4439430    0.2581769
#> [1231]    0.1409285    0.6568281    0.4252400    0.5805476    1.0801982
#> [1236]    0.5515434    0.3120746    0.2868519    0.2142312    0.1409285
#> [1241]    0.1473794    0.3661773    2.3025114    1.4997795    0.1473794
#> [1246]    0.3172646    0.2019896    1.1775833    0.4761386    0.1714962
#> [1251]    0.5732166    0.4502794    0.2184090    0.8441193    0.4069938
#> [1256]    0.1409285    0.2400214    0.1170791    0.1227487    0.1227487
#> [1261]    0.1825644    0.1441296    0.2581769    0.1316188    0.1377764
#> [1266]    0.1863516    0.2400214    0.4314235    0.1256566    0.1256566
#> [1271]    0.1256566    0.5515434    0.1714962    1.0907766    0.3018402
#> [1276]    0.3833703    0.1316188    0.1643615    0.2019896    1.9125081
#> [1281]    0.2535660    0.3225033    0.2581769    0.1643615    0.1170791
#> [1286]    0.1679044    0.3718586    0.2312319    0.1227487    0.9383714
#> [1291]    1.0078292    0.2142312    0.4191073    0.2356026    0.2184090
#> [1296]    0.1441296    0.5515434    0.8169697    0.5233888    0.4761386
#> [1301]    0.1473794    0.2771011    0.3018402    0.2312319    0.1506782
#> [1306]    0.1863516    0.2184090    0.3950817    0.2269094    0.1170791
#> [1311]    0.2490031    0.3833703    0.2101020    0.7217952    0.1863516
#> [1316]    0.3225033    0.1714962    0.1825644    0.1506782    0.2967958
#> [1321]    0.1143171    0.6028634    0.1441296    0.1346732    0.2269094
#> [1326]    0.7642095    0.4314235    0.1409285    5.2410062    0.1316188
#> [1331]    0.2490031    1.1775833    0.2101020    0.3439470    0.7815692
#> [1336]    0.8905197    0.1316188    0.4130253    0.1825644    0.1608674
#> [1341]    0.2356026    3.2775205    1.3751156    1.3033466    1.1336940
#> [1346]    0.8811239    0.8441193    0.3605456    0.1643615    0.1170791
#> [1351]    0.1901877    0.1256566    2.7917858    0.5444253    0.1540257
#> [1356]    0.2226350    2.1889503    0.3892010    0.1346732    0.1643615
#> [1361]    0.2535660    0.2269094    0.2722979    0.5953710    0.2771011
#> [1366]    0.2868519    0.6411373    0.2184090    0.2019896    0.3775896
#> [1371]    0.8811239    0.5096275    0.1825644    0.1608674    0.2535660
#> [1376]    0.1788262    0.2675429    0.1377764    0.1170791    0.4252400
#> [1381]    0.1143171    0.2771011    0.1506782    0.2722979    0.1608674
#> [1386]    0.1473794    0.5096275    0.2868519    0.2722979    0.1256566
#> [1391]    0.3277909    0.1286133    0.4960758    0.1473794    0.2269094
#> [1396]    0.1377764    0.2312319    0.2019896    0.1316188    0.1256566
#> [1401]    0.1409285    0.2142312    0.4069938    0.2101020    0.1409285
#> [1406]    0.1227487    0.1714962    0.5373604    8.6666094    0.1473794
#> [1411]    0.4010126    0.3605456    0.2490031    0.5373604    0.1679044
#> [1416]    0.1316188    0.1574222    0.5233888    0.2269094    0.2226350
#> [1421]    0.1940726    0.4252400    0.3494306    0.2581769    0.1940726
#> [1426]    0.2675429    0.2675429    0.1441296    0.1377764    0.3018402
#> [1431]    0.2101020    0.2019896    0.1170791    0.2917997    0.1574222
#> [1436]    0.1377764    0.2060214    0.1574222    0.3225033    0.5444253
#> [1441]    0.1574222    0.1714962    0.5515434    0.4566671    1.1014153
#> [1446]    0.5805476    0.4191073    0.1409285    0.2060214    0.1143171
#> [1451]    0.2819524    0.2400214    0.1679044    1.0384858    0.1540257
#> [1456]    0.6411373    0.5879324    0.5805476    0.2184090    0.1825644
#> [1461]    0.7815692    0.7134802    0.4314235    0.1863516    0.1409285
#> [1466]    0.2535660    0.8350120    0.6489554    0.4893782    0.4695966
#> [1471]    0.2628359    0.2312319    0.1608674    0.7728612    0.8811239
#> [1476]    0.1643615    1.1228738    0.3331272    0.1788262    0.1227487
#> [1481]    0.1377764    0.3385126    0.2535660    0.1574222    0.4893782
#> [1486]    0.1940726    0.3833703    0.1574222    0.1441296    0.4439430
#> [1491]    0.1441296    0.1198895    0.3950817    0.7556143    0.2535660
#> [1496]    0.1643615    0.2581769    0.2269094    0.5515434    0.1256566
#> [1501]    0.1788262    0.2581769    0.1316188    0.2819524    0.2722979
#> [1506]    0.7815692    0.5732166    0.1714962    0.9579217    0.2675429
#> [1511]    0.2675429    0.2490031    0.1377764    0.4130253    0.3225033
#> [1516]    0.3833703    0.9579217    0.2771011    0.5953710    0.6647558
#> [1521]    0.1679044    0.2184090    0.2269094    0.1714962    0.2400214
#> [1526]    0.2400214    0.1788262    0.3494306    0.1980066    0.1751367
#> [1531]    6.7307993    0.1143171    0.1980066    0.1980066    0.2226350
#> [1536]    0.3661773    0.2535660    0.2967958    0.2184090    0.1346732
#> [1541]    0.4960758    0.2819524    0.1679044    0.1714962    0.1714962
#> [1546]    0.1377764    0.2184090    0.5373604    0.4010126    0.4252400
#> [1551]    0.2581769    0.4130253    0.3718586    0.1901877    0.6256648
#> [1556]    0.1863516    0.1788262    0.2819524    0.6970172    0.1679044
#> [1561]    0.2628359    0.1506782    0.7470753    0.2535660    0.1643615
#> [1566]    0.1540257    0.3277909    0.2356026    0.7903341    0.3277909
#> [1571]    0.1506782    0.2722979    8.1287506    0.1825644    0.3605456
#> [1576]    0.1940726    0.2019896    0.8441193    0.9579217    0.2535660
#> [1581]    1.1998947    0.1788262    0.5732166    0.2967958    0.3439470
#> [1586]    0.1825644    0.2581769    0.1143171    0.2226350    0.1316188
#> [1591]    0.2917997    0.5732166    0.4566671    0.1643615    0.2269094
#> [1596]    0.4314235    0.2675429    0.1227487    0.2868519    0.2184090
#> [1601]    0.2019896    0.3385126    0.3018402    0.1788262    0.4439430
#> [1606]    0.2019896    0.5515434    0.1901877   20.7891965   14.4664031
#> [1611]    0.1286133    0.1751367    0.1901877    0.1679044    0.1227487
#> [1616]    0.5096275    0.1286133    0.2226350    0.6333738    0.3775896
#> [1621]    0.2917997    0.3775896    0.1980066    0.2628359    0.6256648
#> [1626]    0.2868519    0.1540257    0.2535660    0.1679044    0.2101020
#> [1631]    0.1473794    0.2101020    0.4439430    0.2868519    0.4376578
#> [1636]    0.1940726    0.1377764    0.1198895    0.7991557    0.1409285
#> [1641]    0.1198895    0.5515434    0.1316188    0.8169697    0.3549634
#> [1646]    0.1540257    0.1346732    0.1409285    0.1441296    0.1377764
#> [1651]    0.1227487    0.1198895    0.1643615    0.1409285    0.3385126
#> [1656]    0.4439430    0.2868519    0.1863516    0.1346732    0.5164819
#> [1661]    0.5732166    0.4502794    0.1409285    0.5732166    0.1825644
#> [1666]    0.1714962    0.4695966    0.1714962    0.3549634    0.1198895
#> [1671]    0.1227487    0.1316188    0.1286133    0.1940726    0.1506782
#> [1676]    0.3605456    0.1506782    0.1540257    0.4130253    0.1409285
#> [1681]    0.4893782    0.2581769    0.3661773    0.1170791    0.1143171
#> [1686]    0.1346732    0.3120746    0.2917997    5.5156022    0.1506782
#> [1691]    0.2101020    0.1227487    0.2535660    0.1788262    0.4130253
#> [1696]    0.2400214    0.1679044    0.1316188    6.3297584    0.3069331
#> [1701]    0.1316188    0.5953710    0.1863516    0.3069331    0.1256566
#> [1706]    0.6104097    0.3385126    0.6568281    0.1506782    0.2226350
#> [1711]    0.2142312    0.3069331    0.1825644    0.1643615    0.2060214
#> [1716]    0.1198895    0.1227487    0.1863516    0.1940726    0.3172646
#> [1721]    0.2101020    0.2400214    0.1170791    0.1751367    0.1143171
#> [1726]    0.3833703    0.1198895    0.1506782    0.4314235    0.1143171
#> [1731]    0.4191073    0.3225033    0.3439470    0.6180101    0.1198895
#> [1736]    0.1286133    0.2142312    0.1751367    0.3950817    0.2967958
#> [1741]    0.2581769    0.2444883    8.1992767    1.3751156    0.2400214
#> [1746]    0.2868519    0.5444253    0.1473794    0.3331272    0.2581769
#> [1751]    0.1198895    0.3225033    0.1170791    0.6104097    0.2101020
#> [1756]    0.2184090    0.1170791    0.1256566    0.1316188    0.4827324
#> [1761]    0.1377764    0.1473794    0.2356026    0.2101020    1.5514302
#> [1766]    1.1121144    0.1377764    0.1863516    0.2675429    0.1863516
#> [1771]    0.2917997    0.4439430    0.1143171    0.3069331    0.2356026
#> [1776]    0.1679044    0.9094855    0.1227487    0.1506782    0.1346732
#> [1781]    0.2722979    0.5515434    0.2060214    0.7556143    0.3549634
#> [1786]    0.1409285    0.6180101    0.2142312    0.2060214    0.2722979
#> [1791]    0.1377764    0.2771011    0.1170791    0.1825644    0.4893782
#> [1796]    0.1643615    0.1540257    0.1409285    0.3661773    0.4376578
#> [1801]    0.3775896    5.1602663    0.2581769    1.7264324    0.1679044
#> [1806]    0.9481172    0.4069938    0.4069938    0.1980066    0.1980066
#> [1811]    0.2819524    0.2356026    0.2019896    0.3172646    0.1316188
#> [1816]    0.6104097    0.1788262    0.2819524    0.1473794    0.2184090
#> [1821]    0.3549634    0.2581769    0.3494306    0.1346732    0.1679044
#> [1826]    0.1506782    0.7728612    0.1198895    0.2628359    0.2535660
#> [1831]    0.5303482    0.9876887    0.1901877    0.1170791    3.1765989
#> [1836]    0.1788262    0.1473794    0.1170791    0.1377764    0.1409285
#> [1841]    0.1227487    0.1346732    0.1863516    0.7385925    0.1643615
#> [1846]    0.5444253    0.2400214    0.2535660    1.2799237    0.1473794
#> [1851]    0.1143171    0.1316188    0.1286133    0.7728612    0.1506782
#> [1856]    0.1788262    0.1316188    0.1409285    0.5515434    0.5732166
#> [1861]    0.8905197    0.4502794    0.3069331    0.1409285    0.1198895
#> [1866]    0.3718586    0.2535660    0.1643615    3.4429782    0.1863516
#> [1871]    0.1409285    0.3605456    0.2444883    0.1540257    0.3331272
#> [1876]    0.3775896    0.1286133    0.1316188    0.1473794    0.1316188
#> [1881]    0.1473794    0.1198895    0.1346732    0.2226350    0.2226350
#> [1886]    0.2019896    0.9481172    4.0151216    0.2060214    0.2722979
#> [1891]    0.2771011    0.1473794    2.7183992    0.1788262    0.2444883
#> [1896]    0.3494306    0.3069331    0.1286133   42.9870983 2124.6053414
#> [1901]    6.1566643  178.0920750  846.3309709  321.4745009   65.7272787
#> [1906] 2947.4395778 4239.1762738  360.8465310 4455.2576569 3343.8077394
#> [1911] 1168.1579275 2942.8205078  864.9962862  479.1874580   15.9166250
#> [1916]    0.8595952  185.5495835    0.3987035   57.7002111  334.8456118
#> [1921]    1.8713912  274.6279943   11.0367446  166.0969414   26.3446037
#> [1926]    2.9830569 1257.8041227  362.8276720    0.2771011   12.0258147
#> [1931]  882.7238379  810.5606026    0.2273446  121.8995630    0.5723998
#> [1936]   10.1924490  558.6397145   81.2738636  136.6002600    1.6911843
#> [1941]    6.6041731    7.2380649   26.1958155 1228.8316157  542.4862337
#> [1946] 1427.1907704    0.2034647   22.0255422 4561.2880071    0.5557262
#> [1951] 3814.5565857  393.3488395  204.1075572   25.8313766  168.0360673
#> [1956]    0.4035750 1312.6704264  733.2185904  409.5559670 2405.6275764
#> [1961]  641.1324549  127.8821029   15.7631838    2.2407640  986.7512129
#> [1966] 1733.3961280 2166.1697213  568.4497202    0.1800597    7.9773238
#> [1971]  337.6863812 1716.0631325 1854.8571960  288.8846213 1435.9596282
#> [1976]  196.2044468  711.7204971 2236.0050995  545.5091539  509.3615149
#> [1981]  553.0940962 1686.8288554    0.1915546  143.5373542 1550.7121078
#> [1986]    4.0070662  802.2177048  199.9818867 2085.0044868  173.4952599
#> [1991]  994.3146570   23.5163789   91.3053133  962.6368160    5.2063209
#> [1996] 2885.0785047    1.1448304  302.7298377 2569.6713941 6473.8349233
#> [2001]   10.1109423   37.1492770 1031.7997469  137.8108859   49.6654598
#> [2006] 1923.5749033 1053.2715806 1983.3091481   33.6883358    0.1807219
#> [2011] 1127.3206685    1.1336754   61.1776709   47.8340431  389.9155628
#> [2016]    3.2910015    0.3818237  258.5408711    9.3122714 1783.3236923
#> [2021]  533.5858708   58.3445739 3255.9482230 1734.2833098 3209.0523049
#> [2026]  118.8859502    0.4420467    0.3857866    0.3857866    0.7551725
#> [2031]    0.3565246    0.3122661    0.4272509    1.6396560    2.1207698
#> [2036]    0.2552849    0.6800809    0.3122661    1.0651287    0.9971716
#> [2041]    0.3486363    0.2499422    0.2776729    0.7473986    0.3262092
#> [2046]    0.3055835    0.7244310    0.3055835    0.2866096    0.4949337
#> [2051]    0.8964513    0.7244310    0.3023117    0.2662468    0.4521837
#> [2056]    0.3857866    0.3262092    0.2499422    0.6800809    0.3023117
#> [2061]    0.5352578    0.3226494    1.2428261    0.3055835    0.2747669
#> [2066]    0.3486363    0.2927632    0.2776729    0.2634671    0.3371923
#> [2071]    0.2579880    0.4224279    1.4386532    0.3901789    0.3089013
#> [2076]    0.3901789    0.2607148    0.3946247    0.6728937    0.5976530
#> [2081]    0.7094133    0.9413183    1.7192102    0.9783091    0.2552849
#> [2086]    0.4949337    0.3156785    0.6447255    0.2806144    0.2896659
#> [2091]    0.2927632    0.6174822    0.3646216    0.3486363    0.3525544
#> [2096]    0.2690555    0.4321284    0.3023117    0.3023117    0.2866096
#> [2101]    0.5720141    0.3409555    0.2896659    0.2579880    0.3687489
#> [2106]    0.3565246    0.2662468    0.2552849    0.9689697    0.5657465
#> [2111]    2.0603028    0.3525544    0.3409555    0.3226494    0.9142158
#> [2116]    0.2896659    0.2866096    0.3055835    0.2747669    1.0356343
#> [2121]    0.3991240    0.3226494    0.3262092    0.2835928    0.2607148
#> [2126]    0.3191394    0.3447701    0.3409555    0.3122661    0.2959026
#> [2131]    0.3298194    0.3055835    0.5352578    0.3055835    0.6174822
#> [2136]    0.3946247    0.5720141    0.5783384    0.2866096    1.2428261
#> [2141]    0.3298194    0.3729289    0.7949313    0.2776729    0.2747669
#> [2146]    0.2806144    0.3447701    0.6657646    0.2959026    0.6516806
#> [2151]    0.3191394    0.2776729    0.2634671    0.2806144    0.5533816
#> [2156]    0.3122661    0.2806144    0.5005266    0.3191394    0.5352578
#> [2161]    0.6042054    0.4224279    0.4784878    0.3023117    0.2526036
#> [2166]    0.3298194    0.3565246    0.4731167    0.4839143    0.9322234
#> [2171]    0.6309885    0.3262092    0.3409555    0.3771617    0.3687489
#> [2176]    0.3447701    0.3371923    0.3328332    0.3923792    0.2134428
#> [2181]    2.3732591    0.2525569    0.2366605    0.2321704    0.3001404
#> [2186]    0.2797214    0.2496457    0.2958527    0.2177047    0.2134428
#> [2191]    0.2121244    0.4845566    0.2496457    0.2208538    0.2148087
#> [2196]    0.2134428    0.2835987    0.2261322    0.3090116    0.2134428
#> [2201]    0.2835987    0.2148087    0.1010070    1.6514341    1.5852423
#> [2206]    1.4204132    0.8714379    1.3597798    0.1705048    0.8901397
#> [2211]    0.1321909    0.1689780    0.4191428    0.2802842    0.4971532
#> [2216]    0.1653795    0.2515981    0.2246966    0.1763274    0.3375116
#> [2221]    0.1203353    0.1800785    0.3712776    0.2609622    0.1261621
#> [2226]    0.3107529    0.2203874    0.1448566    0.1416140    0.1321909
#> [2231]   92.0622839 1358.5939715   24.5657667    4.1871881    0.2304051
#> [2236]    0.2260164    2.3060455  517.0387765    2.5306817    6.7888058
#> [2241]    0.1771949    1.2203708    0.1966709   30.0602730    2.3738264
#> [2246]   30.4540861   51.0408111   58.5667103    0.1697535    0.1848358
#> [2251]   88.6653382    2.2559820    0.1293140    3.5588924    0.4467054
#> [2256]    5.7751178   86.1404237   55.6860176   49.8640722    2.5843162
#> [2261]    8.6645655  158.1721940   41.9859219    9.1961508  131.1407278
#> [2266]    2.6568882    3.8126577   15.1706384    4.4558352    1.0811392
#> [2271]   10.8880738    0.6746019    0.2919361    5.1331512    0.2563600
#> [2276]    0.7658518    0.2121546    0.3062288    1.1220370    3.7449062
#> [2281]    0.5849885   11.4927141    1.4998599   14.1074636    0.4085149
#> [2286]    0.5164652    0.2884804

# split dataset to avoid memory over usage
data_split = split(scbi_stem1, cut(1:nrow(scbi_stem1), breaks = 10, labels = FALSE))
agb = lapply(data_split, function(df) get_biomass(dbh=df$dbh,
                                                 genus=df$genus,
                                                 species = df$species,
                                                 coords = c(-78.2, 38.9)))
scbi_stem1$agb = do.call(c, agb)