Join allodb tables: equations, sitespecies, and sites_info.

master() has all column types set as characters to preserve different kinds of missing values.

master_tidy() has column types and missing optimized for computation.

master()

master_tidy()

Value

A dataframe.

See also

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(purrr)

master()

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> Error: Join columns must be present in data.
#> x Problem with `equation_id`.

# Columns are of type "character"
master() %>%
  purrr::map_chr(typeof) %>%
  unique()

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> Error: Join columns must be present in data.
#> x Problem with `equation_id`.

# This preserves preserve different kinds of NAs
unique(grep("N", master()$dbh_min_cm, value = TRUE))

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> Error: Join columns must be present in data.
#> x Problem with `equation_id`.

allodb::missing_values_metadata

#> # A tibble: 4 x 3
#>   Code  Definition         Description                                          
#>   <chr> <chr>              <chr>                                                
#> 1 NA    Not Applicable     Data does not apply to that particular case          
#> 2 NAC   Not Acquired       Information may be available but has not been acquir~
#> 3 NRA   Not Readily Avail~ Information was not readily available to the authors~
#> 4 NI    No Information     No information available in original publication     

# For computation, use `master_tidy()`, which has column types and missing
# values already optimized for computation on the data
master_tidy() %>%
  map_chr(typeof) %>%
  unique()

#> Joining `equations` and `sitespecies` by 'equation_id'; then `sites_info` by 'site'.

#> Error: Join columns must be present in data.
#> x Problem with `equation_id`.