Create objects of class allodb.

This function creates S3 objects of class "allodb". Compared with the data directly delivered in the allodb package, objects of class "allodb" have two main differences:

The column types are not all text but the types that is most useful for computations.

Missing values are represented in a single way: with NA.

set_type(x)

Arguments

x
A dataframe from allodb.

Value

A dataframe subclass "allodb".

Details

The datasets directly delivered by allodb are optimized for delivering the greatest amount of information in the most robust and flexible way. All values are stored as text. This allows, for example, different to store different representation of missing values. The trade-off is that text values don't allow direct computation. First they may need to be converted to other types of data, such as integers.

See also

Examples

set_type(allodb::equations)

#> # A tibble: 567 x 47
#>    ref_id equation_id equation_allome~ equation_form dependent_varia~
#>  * <chr>  <chr>       <chr>            <chr>         <chr>           
#>  1 baske~ e2c7c7      exp(0.15+2.48*l~ exp(a+b*(log~ Total abovegrou~
#>  2 bohn_~ 307ec5      exp((1.185*(log~ exp((a*(dbh-~ Total abovegrou~
#>  3 bohn_~ 6a7382      exp((1.151*(log~ exp((a*(dbh-~ Total abovegrou~
#>  4 bohn_~ 786b7c      exp((1.266*(log~ exp((a*(dbh-~ Total abovegrou~
#>  5 bohn_~ 9e3b6a      exp((1.192*(log~ exp((a*(dbh-~ Total abovegrou~
#>  6 bohn_~ 9f138f      exp((1.202*(log~ exp((a*(dbh-~ Total abovegrou~
#>  7 bohn_~ e42e41      exp((1.091*(log~ exp((a*(dbh-~ Total abovegrou~
#>  8 bohn_~ e57b77      exp((1.029*(log~ exp((a*(dbh-~ Total abovegrou~
#>  9 bohn_~ f96447      dbh/((1/1.879)+~ dbh((1/a)+(d~ Height          
#> 10 bohn_~ 7a3dce      dbh/((1/1.919)+~ dbh((1/a)+(d~ Height          
#> # ... with 557 more rows, and 42 more variables: independent_variable <chr>,
#> #   equation_taxa <chr>, allometry_specificity <chr>, equation_categ <chr>,
#> #   geographic_area <chr>, original_coord <chr>, lat <dbl>, long <dbl>,
#> #   elev_m <chr>, `geography. Notes` <chr>, mat_C <chr>, min.temp_C <dbl>,
#> #   max.temp_C <dbl>, map_mm <chr>, frost_free_period_days <chr>, `climate.
#> #   Notes` <chr>, stand_age_range_yr <chr>, stand_age_history <chr>,
#> #   stand_basal_area_m2_ha <chr>, stand_trees_ha <chr>,
#> #   forest_description <chr>, ecosystem_type <chr>, koppen <chr>,
#> #   dbh_min_cm <dbl>, dbh_max_cm <dbl>, sample_size <int>,
#> #   collection_year <chr>, dbh_units_original <chr>, dbh_unit_CF <dbl>,
#> #   output_units_original <chr>, output_units_CF <dbl>,
#> #   allometry_development_method <chr>, regression_model <chr>,
#> #   r_squared <dbl>, other_equations_tested <chr>, log_biomass <chr>,
#> #   bias_corrected <chr>, bias_correction_factor <chr>,
#> #   notes_fitting_model <chr>, original_equation_id <chr>,
#> #   original_data_availability <chr>, equation_notes <chr>

class(set_type(allodb::equations))

#> [1] "allodb"      "spec_tbl_df" "tbl_df"      "tbl"         "data.frame"